% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{all_cell_ids}
\alias{all_cell_ids}
\title{Get all cell IDs in store}
\usage{
all_cell_ids(store)
}
\arguments{
\item{store}{A cell_store object}
}
\value{
Character vector of all cell IDs
}
\description{
Get all cell IDs in store
}
\keyword{internal}
