% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{any_grepl}
\alias{any_grepl}
\title{any_grepl}
\usage{
any_grepl(
  pattern_vector,
  find_in_vector,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{pattern_vector}{the vector of strings to look for}

\item{find_in_vector}{the vector of strings to look in}

\item{ignore.case}{per base grepl}

\item{perl}{per base grepl}

\item{fixed}{per base grepl}

\item{useBytes}{per base grepl}
}
\value{
a boolean vector
}
\description{
grepl wrapper for any matches in any vector of patterns.
allows spaces, unlike passing e.g. "pattern1 | pattern2" to grepl directly
}
\examples{
any_grepl(c("cat", "shoe"), c("shoe dog", "hat dog", "cat dog"))
}
