% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{any_gsub}
\alias{any_gsub}
\title{any_gsub}
\usage{
any_gsub(
  pattern_vector,
  replacement_value = "",
  find_in_vector,
  trimws_at_end = TRUE,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{pattern_vector}{the vector of strings to look for}

\item{replacement_value}{the string to replace with}

\item{find_in_vector}{the vector of strings to look in}

\item{trimws_at_end}{whether to trim white space}

\item{ignore.case}{per base gsub}

\item{perl}{per base gsub}

\item{fixed}{per base gsub}

\item{useBytes}{per base gsub}
}
\value{
find_in_vector with replacements made
}
\description{
gsub wrapper to replace all matches to a vector of strings with a single replacement value
uses fixed = TRUE in gsub and trims whitespace at the end by default
}
\examples{
any_gsub(c("cat", "hat"), c("shoe dog", "hat dog", "cat dog"), "")
}
