% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{append_oe_to_oe_duplicates}
\alias{append_oe_to_oe_duplicates}
\title{Append "oe" to Variable Labels for Open-Ended Duplicates}
\usage{
append_oe_to_oe_duplicates(temp_dat, temp_dpdict)
}
\arguments{
\item{temp_dat}{A survey data dataframe}

\item{temp_dpdict}{A dpdict dataframe with variable metadata including variable_labels}
}
\value{
The updated dpdict with "oe" appended to variable labels for open-ended
duplicates. Variable labels are modified from format "Q1 - Statement" to
"Q1oe - Statement" for affected variables.
}
\description{
Identifies character variables in survey data that have duplicate variable labels
with their non-open-ended counterparts and appends "oe" to the variable name prefix
in their labels to distinguish them.
}
\examples{
\dontrun{
# Assumes temp_dat has character variables with duplicate labels
temp_dat <- data.frame(
  Q1_1 = c(1, 2, 3),
  Q1_1_oe = c("Other text", "More text", "Additional text")
)
attr(temp_dat$Q1_1, "label") <- "Q1 - Brand preference"
attr(temp_dat$Q1_1_oe, "label") <- "Q1 - Brand preference"

temp_dpdict <- create_dict(temp_dat)
updated_dpdict <- append_oe_to_oe_duplicates(temp_dat, temp_dpdict)
}
}
