% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{as.character.formula_with_tolerance}
\alias{as.character.formula_with_tolerance}
\title{Convert formula_with_tolerance to character}
\usage{
\method{as.character}{formula_with_tolerance}(x, ...)
}
\arguments{
\item{x}{A formula_with_tolerance object}

\item{...}{Additional arguments (ignored)}
}
\description{
Convert formula_with_tolerance to character
}
