% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{base_calculators}
\alias{base_calculators}
\alias{base_column_total}
\alias{base_row_total}
\alias{base_cell_count}
\alias{base_grand_total}
\alias{base_column_total_valid}
\alias{base_row_total_valid}
\alias{base_cell_count_valid}
\title{Export base calculators for user access}
\usage{
base_column_total(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_row_total(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_cell_count(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_grand_total(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_column_total_valid(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_row_total_valid(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)

base_cell_count_valid(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)
}
\description{
Export base calculators for user access
}
