% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{base_cell_count}
\alias{base_cell_count}
\title{Built-in base calculator: Cell count}
\usage{
base_cell_count(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  col_filter = NULL,
  row_filter = NULL,
  ...
)
}
\description{
Base is the count in the specific cell
}
\keyword{internal}
