% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{base_cell_count_valid}
\alias{base_cell_count_valid}
\title{Built-in base calculator: Cell count excluding NA values}
\usage{
base_cell_count_valid(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  ...
)
}
\description{
Base is the cell count excluding NA in values array
}
\keyword{internal}
