% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{base_column_total}
\alias{base_column_total}
\title{Built-in base calculator: Column total}
\usage{
base_column_total(base_array, row_array, col_array, values_array = NULL, ...)
}
\description{
Base is the total count in the column
}
\keyword{internal}
