% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{base_column_total_valid}
\alias{base_column_total_valid}
\title{Built-in base calculator: Column total excluding NA values}
\usage{
base_column_total_valid(
  base_array,
  row_array,
  col_array,
  values_array = NULL,
  ...
)
}
\description{
Base is the column total excluding NA in values array
}
\keyword{internal}
