% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{base_grand_total}
\alias{base_grand_total}
\title{Built-in base calculator: Grand total}
\usage{
base_grand_total(base_array, row_array, col_array, values_array = NULL, ...)
}
\description{
Base is the total count across all data
}
\keyword{internal}
