% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{base_row_total}
\alias{base_row_total}
\title{Built-in base calculator: Row total}
\usage{
base_row_total(base_array, row_array, col_array, values_array = NULL, ...)
}
\description{
Base is the total count in the row
}
\keyword{internal}
