% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{bottom_box}
\alias{bottom_box}
\title{Select bottom n response options}
\usage{
bottom_box(var, n = 1)
}
\arguments{
\item{var}{Variable name or expression}

\item{n}{Number of bottom options to select (default 1)}
}
\value{
Expression for use in tab()
}
\description{
Select bottom n response options
}
