% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{build_nary_call}
\alias{build_nary_call}
\title{Build N-ary Call}
\usage{
build_nary_call(op, args)
}
\arguments{
\item{op}{Operator}

\item{args}{List of arguments}
}
\value{
Nested call expression
}
\description{
Builds nested binary calls from list: list(a, b, c) with & → a & b & c
}
\keyword{internal}
