% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{calculate_summary}
\alias{calculate_summary}
\title{Calculate summary value based on type}
\usage{
calculate_summary(type, arrays, base_array, col_array, statistic, values)
}
\description{
Calculate summary value based on type
}
\keyword{internal}
