% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{$.tab_result}
\alias{$.tab_result}
\title{Extract element from tab_result using $}
\usage{
\method{$}{tab_result}(x, name)
}
\arguments{
\item{x}{A tab_result object}

\item{name}{Element name}
}
\description{
Extract element from tab_result using $
}
