% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{cell_count}
\alias{cell_count}
\title{Get number of cells in store}
\usage{
cell_count(store)
}
\arguments{
\item{store}{A cell_store object}
}
\value{
Integer count of cells
}
\description{
Get number of cells in store
}
\keyword{internal}
