% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{change_from}
\alias{change_from}
\title{Calculate change from baseline}
\usage{
change_from(var, condition)
}
\arguments{
\item{var}{Variable name or expression}

\item{condition}{Condition defining baseline}
}
\value{
Expression for use in tab()
}
\description{
Calculate change from baseline
}
