% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{check_dat_dpdict_alignment}
\alias{check_dat_dpdict_alignment}
\title{check_dat_dpdict_alignment
simple checks that dat and dpdict are in a suitable and aligned format
by default, prints report of results}
\usage{
check_dat_dpdict_alignment(temp_dat, temp_dpdict, report = TRUE)
}
\arguments{
\item{temp_dat}{survey data dataframe}

\item{temp_dpdict}{'dpdict' to check for alignment with temp_dat}

\item{report}{whether to print a report or not}
}
\value{
TRUE if passes all checks
}
\description{
check_dat_dpdict_alignment
simple checks that dat and dpdict are in a suitable and aligned format
by default, prints report of results
}
\examples{
temp_dat <- get_minimal_labelled_test_dat()
temp_dpdict <- create_dict(temp_dat)
check_dat_dpdict_alignment(temp_dat, temp_dpdict)
}
