% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{check_for_common_alphabets}
\alias{check_for_common_alphabets}
\title{check_for_common_alphabets}
\usage{
check_for_common_alphabets(input_string)
}
\arguments{
\item{input_string}{a string}
}
\value{
a comma delineated string of common alphabet names recognised in the input_string
}
\description{
returns a comma delineated string of common alphabet names recognised in the input_string
looks for Latin, Cyrillic, Arabic, CJK, Devanagari and Bengali
}
\examples{
\donttest{
check_for_common_alphabets("Latin, Русский, عربي, 漢語, বাংলা, देवनागरी")
}
}
