% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{check_key_consistency}
\alias{check_key_consistency}
\title{check_key_consistency}
\usage{
check_key_consistency(temp_dat, key_name)
}
\arguments{
\item{temp_dat}{dataframe to check in}

\item{key_name}{unique key}
}
\value{
TRUE if passes consistency checks, else returns a data frame to help diagnose inconsistencies
}
\description{
checks that values in all other columns for every unique value of key_name are the same for every instance of that value of key_name
if false, returns a data frame to help diagnose inconsistencies
note: doesn't check uniqueness of key_name, i.e. there can be multiple (duplicate) instances of the same value of key_name
}
\examples{
one_inconsistent_df <- data.frame(id = c(1, 1, 2, 2, 3),
  value1 = c("A", "A", "B", "C", "D"),
  value2 = c(10, 10, 20, 20, 30))
concatenate_by_group(one_inconsistent_df, "id")
}
