% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_variables_exist}
\alias{check_variables_exist}
\title{Check if all variables exist in data and provide helpful error message}
\usage{
check_variables_exist(vars, data, context = "expression")
}
\arguments{
\item{vars}{Character vector of variable names to check}

\item{data}{Data frame to check against}

\item{context}{String describing where the variables are being used}
}
\description{
Check if all variables exist in data and provide helpful error message
}
\keyword{internal}
