% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{clear_tab_registry}
\alias{clear_tab_registry}
\title{Clear all registered helpers and statistics (useful for testing)}
\usage{
clear_tab_registry()
}
\description{
Clear all registered helpers and statistics (useful for testing)
}
