% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{closest_matching_string}
\alias{closest_matching_string}
\title{closest_matching_string}
\usage{
closest_matching_string(strings_vector, string_to_check, method = "lv")
}
\arguments{
\item{strings_vector}{vector of strings}

\item{string_to_check}{single string to measure distance against}

\item{method}{optionally specify a distance method. method = lv by default.}
}
\value{
string from strings_vector that best matches string_to_check
}
\description{
returns closest matching string from strings_vector to string_to_check
}
\examples{
closest_matching_string(c("shoe", "dog", "cat"), "dug")
}
