% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{col_def_matches_expression}
\alias{col_def_matches_expression}
\title{Check if col_def matches an expression}
\usage{
col_def_matches_expression(col_def, expr)
}
\arguments{
\item{col_def}{A layout_def object}

\item{expr}{Expression to check}
}
\value{
Logical
}
\description{
Check if col_def matches an expression
}
\keyword{internal}
