% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{colnames.tab_result}
\alias{colnames.tab_result}
\title{Get column names of tab_result}
\usage{
colnames.tab_result(x)
}
\arguments{
\item{x}{A tab_result object}
}
\description{
Get column names of tab_result
}
