% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{compute_cell}
\alias{compute_cell}
\title{Compute value for a single cell}
\usage{
compute_cell(
  base_array,
  row_array,
  col_array,
  statistic = "column_pct",
  values = NULL
)
}
\arguments{
\item{row_array}{Numeric array for row specification}

\item{col_array}{Numeric array for column specification (NULL for marginals)}

\item{statistic}{Type of statistic to compute}

\item{values}{Optional values array for mean calculations}
}
\value{
Computed cell value
}
\description{
Compute value for a single cell
}
\keyword{internal}
