% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{compute_cells_vectorized}
\alias{compute_cells_vectorized}
\title{Compute values for multiple cells using vectorized operations}
\usage{
compute_cells_vectorized(
  base_array,
  row_arrays,
  col_arrays,
  row_u_arrays = NULL,
  col_u_arrays = NULL,
  statistic = "column_pct",
  values = NULL
)
}
\arguments{
\item{base_array}{Base array for filtering}

\item{row_arrays}{List of row arrays}

\item{col_arrays}{List of column arrays}

\item{statistic}{Type of statistic to compute}

\item{values}{Optional values array for mean calculations}
}
\value{
Matrix of computed cell values
}
\description{
Compute values for multiple cells using vectorized operations
}
\keyword{internal}
