% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{compute_significance}
\alias{compute_significance}
\title{Compute significance matrix for tab results}
\usage{
compute_significance(
  base_array,
  row_arrays,
  col_arrays,
  result_matrix,
  statistic,
  values_array = NULL,
  sig_config = list(),
  row_labels = NULL
)
}
\description{
Compute significance matrix for tab results
}
\keyword{internal}
