% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{concatenate_by_group}
\alias{concatenate_by_group}
\title{concatenate_by_group}
\usage{
concatenate_by_group(temp_dat, group_var)
}
\arguments{
\item{temp_dat}{a dataframe}

\item{group_var}{the name of the grouping variable}
}
\value{
the grouped dataframe with concatenated values
}
\description{
returns a data frame grouped by group_var where in case of multiple rows within a group, the value for that group is a concatenation of unique values for those rows
}
\examples{
temp_dat <- dplyr::tibble(
  ID = c(1, 1, 2, 2, 3),
  Name = c("John", "John", "Alice", "Alice", "Bob"),
  City = c("New York", "Los Angeles", "Chicago", "Chicago", "Houston"),
  Age = c(25, 25, 30, 30, 35)
)
concatenate_by_group(temp_dat, "ID")
}
