% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_custom_col_def}
\alias{create_custom_col_def}
\title{Create custom column definition}
\usage{
create_custom_col_def(fn, label)
}
\arguments{
\item{fn}{Custom function taking a cell and returning TRUE/FALSE}

\item{label}{Label for the position}
}
\value{
layout_def object
}
\description{
Create custom column definition
}
\keyword{internal}
