% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{create_derive_operation}
\alias{create_derive_operation}
\title{Create and register a derive operation}
\usage{
create_derive_operation(id, processor, description = NULL)
}
\arguments{
\item{id}{Character identifier for the operation}

\item{processor}{Function that processes the operation
Signature: function(tab_result, ...)}

\item{description}{Optional description of what this operation does}
}
\value{
The created operation object (invisibly)
}
\description{
Create and register a derive operation
}
