% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{create_groups_from_values}
\alias{create_groups_from_values}
\title{Create groups from specific values of a variable}
\usage{
create_groups_from_values(var_name, values, data, dpdict = NULL)
}
\description{
Create groups from specific values of a variable
}
\keyword{internal}
