% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{create_groups_from_variable}
\alias{create_groups_from_variable}
\title{Create groups from all unique values of a variable}
\usage{
create_groups_from_variable(var_name, data, dpdict = NULL)
}
\description{
Create groups from all unique values of a variable
}
\keyword{internal}
