% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{create_helper}
\alias{create_helper}
\title{Create and register a custom helper}
\usage{
create_helper(id, processor, ...)
}
\arguments{
\item{id}{Character identifier for the helper}

\item{processor}{Function that processes the helper specification}

\item{...}{Additional attributes}
}
\value{
The created helper object (invisibly)
}
\description{
Create and register a custom helper
}
