% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_significance_test}
\alias{create_significance_test}
\title{Create and register a significance test}
\usage{
create_significance_test(id, processor, name = NULL, description = NULL, ...)
}
\arguments{
\item{id}{Character identifier for the test}

\item{processor}{Function that performs the test}

\item{name}{Human-readable name for the test}

\item{description}{Description of when to use this test}

\item{...}{Additional attributes}
}
\value{
The created test object (invisibly)
}
\description{
Create and register a significance test
}
