% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{datamap}
\alias{datamap}
\title{datamap
creates a dataframe with variable names, labels and other useful info as rows (SPSS style)
useful for quickly understanding structure of a large dataset
note: order of value labels and order of unique values does not necessarily correspond}
\usage{
datamap(temp_dat, view_or_return = "view")
}
\arguments{
\item{temp_dat}{a dataframe, assumed to be survey data}

\item{view_or_return}{calls View by default - or if "save", return the summary as a dataframe}
}
\value{
calls View by default - or if "return", return the summary
}
\description{
datamap
creates a dataframe with variable names, labels and other useful info as rows (SPSS style)
useful for quickly understanding structure of a large dataset
note: order of value labels and order of unique values does not necessarily correspond
}
\examples{
datamap(get_basic_test_dat(100), view_or_return = "return")
}
