% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{datamap_questions}
\alias{datamap_questions}
\title{datamap_questions}
\usage{
datamap_questions(survey_obj)
}
\arguments{
\item{survey_obj}{A survey_data object containing the survey data and metadata.}
}
\value{
a data frame with columns for question_group, var_count, first_var_label, first_var_name, var_classes, var_values, var_labels, and question_types
}
\description{
provides a convenient view similar to SPSS variables like datamap, but by question group instead of individual variables.
}
\examples{
datamap_questions(create_survey_data(get_big_test_dat(10)))
}
