% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Main derive function - apply a derive operation to a tab_result}
\usage{
derive(tab_result, operation, ...)
}
\arguments{
\item{tab_result}{A tab_result object from tab()}

\item{operation}{Either a derive operation specification object or a character ID}

\item{...}{Additional arguments passed to the operation processor}
}
\value{
Modified tab_result with derive operation applied
}
\description{
Main derive function - apply a derive operation to a tab_result
}
