% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{detect_dimension_type}
\alias{detect_dimension_type}
\title{Detect whether keys represent variables or values}
\usage{
detect_dimension_type(keys, dpdict)
}
\arguments{
\item{keys}{Vector of keys (character or numeric)}

\item{dpdict}{Data dictionary}
}
\value{
Character: "variable" or "value"
}
\description{
Determines if a set of keys are variable names or value codes by checking
if they match variables in the data dictionary.
}
\keyword{internal}
