% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{dimnames.tab_result}
\alias{dimnames.tab_result}
\title{Get dimension names (row and column names) of tab_result}
\usage{
\method{dimnames}{tab_result}(x)
}
\description{
Get dimension names (row and column names) of tab_result
}
