% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{display_reactable}
\alias{display_reactable}
\title{Display a reactable_tab object as an interactive HTML table}
\usage{
display_reactable(x, height = "auto", width = "100\%", ...)
}
\arguments{
\item{x}{A reactable_tab object created by \code{\link{tab_to_reactable}}}

\item{height}{Table height (default "auto")}

\item{width}{Table width (default "100\%")}

\item{...}{Additional arguments passed to reactable::reactable()}
}
\value{
A reactable widget or htmltools tag list
}
\description{
Display a reactable_tab object as an interactive HTML table
}
\examples{
\dontrun{
r_tab <- data \%>\% tab(q1, gender) \%>\% tab_to_reactable()
display_reactable(r_tab)
}
}
