% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.apply_flourish_settings}
\alias{.apply_flourish_settings}
\title{Apply Flourish settings to widget}
\usage{
.apply_flourish_settings(widget, settings, chart_type)
}
\description{
Apply Flourish settings to widget
}
\keyword{internal}
