% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.apply_percent_scaling}
\alias{.apply_percent_scaling}
\title{Apply percent scaling to numeric columns}
\usage{
.apply_percent_scaling(data, percent_scale, stat_id)
}
\description{
Apply percent scaling to numeric columns
}
\keyword{internal}
