% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.apply_table_flourish_settings}
\alias{.apply_table_flourish_settings}
\title{Apply table settings to Flourish widget using set_table_details}
\usage{
.apply_table_flourish_settings(widget, settings)
}
\description{
Apply table settings to Flourish widget using set_table_details
}
\keyword{internal}
