% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.bind_data_auto}
\alias{.bind_data_auto}
\title{Bind data to Flourish widget based on chart type}
\usage{
.bind_data_auto(widget, ft, chart_type, api_key, data_for_binding = NULL)
}
\description{
Bind data to Flourish widget based on chart type
}
\keyword{internal}
