% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.build_cell_render_fn}
\alias{.build_cell_render_fn}
\title{Build cell rendering function}
\usage{
.build_cell_render_fn(x, col_name)
}
\description{
Build cell rendering function
}
\keyword{internal}
