% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.build_footer_cell_render_fn}
\alias{.build_footer_cell_render_fn}
\title{Build cell rendering function for footer table}
\usage{
.build_footer_cell_render_fn(x, col_name, footer_data)
}
\description{
Build cell rendering function for footer table
}
\keyword{internal}
