% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.build_footer_columns}
\alias{.build_footer_columns}
\title{Build column definitions for footer table}
\usage{
.build_footer_columns(x, footer_data)
}
\description{
Build column definitions for footer table
}
\keyword{internal}
