% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.build_heatmap_color_rules}
\alias{.build_heatmap_color_rules}
\title{Build Flourish color rules for heatmap}
\usage{
.build_heatmap_color_rules(df, ts)
}
\description{
Build Flourish color rules for heatmap
}
\keyword{internal}
