% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.build_reactable_columns}
\alias{.build_reactable_columns}
\title{Build column definitions for reactable}
\usage{
.build_reactable_columns(x, data)
}
\description{
Build column definitions for reactable
}
\keyword{internal}
