% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.build_tooltip_text}
\alias{.build_tooltip_text}
\title{Build tooltip text for a cell}
\usage{
.build_tooltip_text(specification, statistic_name, base_n)
}
\arguments{
\item{specification}{Cell specification list}

\item{statistic_name}{Character name of the statistic}

\item{base_n}{Numeric base count}
}
\value{
Character tooltip text (with HTML)
}
\description{
Constructs a tooltip string describing what the cell represents.
Tailors the message based on the statistic type.
}
\keyword{internal}
