% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.calculate_heatmap_color}
\alias{.calculate_heatmap_color}
\title{Calculate heatmap color for a cell}
\usage{
.calculate_heatmap_color(x, col_name, value, index)
}
\description{
Calculate heatmap color for a cell
}
\keyword{internal}
