% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.clean_reactable_data}
\alias{.clean_reactable_data}
\title{Clean tab data for reactable display}
\usage{
.clean_reactable_data(
  tab_result,
  metadata,
  show_base,
  strip_summary_rows,
  strip_summary_cols
)
}
\description{
Clean tab data for reactable display
}
\keyword{internal}
