% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.clean_tab_data}
\alias{.clean_tab_data}
\title{Clean tab data by removing metadata rows/columns}
\usage{
.clean_tab_data(
  tab_result,
  stat_info,
  strip_base,
  strip_summary_rows,
  strip_summary_cols
)
}
\description{
Clean tab data by removing metadata rows/columns
}
\keyword{internal}
