% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.combine_tables}
\alias{.combine_tables}
\title{Combine main and footer tables with CSS styling}
\usage{
.combine_tables(main_table, footer_table, font_family, font_size)
}
\description{
Combine main and footer tables with CSS styling
}
\keyword{internal}
