% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.create_footer_table}
\alias{.create_footer_table}
\title{Create footer reactable table (base/summary rows only)}
\usage{
.create_footer_table(x, data, width)
}
\description{
Create footer reactable table (base/summary rows only)
}
\keyword{internal}
