% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.create_main_table}
\alias{.create_main_table}
\title{Create main reactable table (data rows only)}
\usage{
.create_main_table(x, data, height, width, ...)
}
\description{
Create main reactable table (data rows only)
}
\keyword{internal}
