% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.create_single_table}
\alias{.create_single_table}
\title{Create single table (when no split needed)}
\usage{
.create_single_table(x, height, width, ...)
}
\description{
Create single table (when no split needed)
}
\keyword{internal}
