% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{.ensure_meta}
\alias{.ensure_meta}
\title{Ensure array has meta attribute with standard structure}
\usage{
.ensure_meta(arr, ivar = NULL, ival = NULL, label = NULL)
}
\arguments{
\item{arr}{Numeric array}

\item{ivar}{Variable label(s)}

\item{ival}{Value label(s)}

\item{label}{Display label}
}
\value{
Array with meta attribute attached
}
\description{
Ensure array has meta attribute with standard structure
}
\keyword{internal}
