% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.extract_reactable_metadata}
\alias{.extract_reactable_metadata}
\title{Extract metadata from tab_result for reactable display}
\usage{
.extract_reactable_metadata(tab_result)
}
\description{
Extract metadata from tab_result for reactable display
}
\keyword{internal}
