% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.extract_stat_info}
\alias{.extract_stat_info}
\title{Extract statistic information from tab result}
\usage{
.extract_stat_info(tab_result)
}
\description{
Extract statistic information from tab result
}
\keyword{internal}
