% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.get_default_settings}
\alias{.get_default_settings}
\title{Generate default Flourish settings based on statistic type}
\usage{
.get_default_settings(stat_id, chart_type)
}
\description{
Generate default Flourish settings based on statistic type
}
\keyword{internal}
