% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.get_default_table_settings}
\alias{.get_default_table_settings}
\title{Get default table settings}
\usage{
.get_default_table_settings()
}
\description{
Get default table settings
}
\keyword{internal}
