% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{.get_hidden_col_indices}
\alias{.get_hidden_col_indices}
\title{Determine which columns are hidden by filter rules}
\usage{
.get_hidden_col_indices(tab_result)
}
\description{
Determine which columns are hidden by filter rules
}
\keyword{internal}
