% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.get_significance_colors}
\alias{.get_significance_colors}
\title{Get default significance colors}
\usage{
.get_significance_colors()
}
\description{
Get default significance colors
}
\keyword{internal}
