% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.get_significance_for_cell}
\alias{.get_significance_for_cell}
\title{Get significance result for a cell}
\usage{
.get_significance_for_cell(significance, sig_comparison, row_label, col_name)
}
\description{
Get significance result for a cell
}
\keyword{internal}
