% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.get_table_flourish_settings}
\alias{.get_table_flourish_settings}
\title{Generate Flourish settings from table_settings}
\usage{
.get_table_flourish_settings(table_settings, df, stat_info)
}
\description{
Generate Flourish settings from table_settings
}
\keyword{internal}
