% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.interpolate_color}
\alias{.interpolate_color}
\title{Interpolate color based on value}
\usage{
.interpolate_color(value, min_val, mid_val, max_val, palette)
}
\description{
Interpolate color based on value
}
\keyword{internal}
