% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.needs_split_display}
\alias{.needs_split_display}
\title{Check if table needs to be split into main + footer}
\usage{
.needs_split_display(x)
}
\description{
Check if table needs to be split into main + footer
}
\keyword{internal}
