% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{.onLoad}
\alias{.onLoad}
\title{Register built-in helpers and statistics}
\usage{
.onLoad(libname = NULL, pkgname = NULL)
}
\arguments{
\item{libname}{Library name (unused)}

\item{pkgname}{Package name (unused)}
}
\description{
Register built-in helpers and statistics
}
\keyword{(unlike}
\keyword{-}
\keyword{ANY}
\keyword{All}
\keyword{Creates}
\keyword{Each}
\keyword{Finds}
\keyword{Match}
\keyword{a}
\keyword{across}
\keyword{against}
\keyword{aggregate}
\keyword{aggregates).}
\keyword{all}
\keyword{and}
\keyword{array}
\keyword{arrays}
\keyword{as}
\keyword{binary}
\keyword{by}
\keyword{each}
\keyword{for}
\keyword{group}
\keyword{have}
\keyword{helper}
\keyword{in}
\keyword{indicating}
\keyword{individual}
\keyword{internal}
\keyword{items}
\keyword{labels}
\keyword{match}
\keyword{matching}
\keyword{multiple}
\keyword{of}
\keyword{one}
\keyword{or}
\keyword{pattern}
\keyword{pattern.}
\keyword{patterns}
\keyword{positive}
\keyword{produces}
\keyword{question}
\keyword{respondents}
\keyword{response}
\keyword{response_match}
\keyword{responses}
\keyword{returns}
\keyword{row}
\keyword{separate}
\keyword{that}
\keyword{value}
\keyword{variable}
\keyword{variables}
\keyword{variables.}
\keyword{which}
\keyword{whose}
\keyword{within}
