% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{.process_banner_arguments}
\alias{.process_banner_arguments}
\title{Process banner helper arguments (internal)}
\usage{
.process_banner_arguments(formula_spec, data)
}
\arguments{
\item{formula_spec}{Formula specification with args}

\item{data}{Data frame for evaluating additional arguments}
}
\value{
List of evaluated arguments
}
\description{
Shared logic for processing banner() special arguments used by both
process_helper and process_helper_for_specs.
}
\keyword{internal}
