% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.safe_bind}
\alias{.safe_bind}
\title{Safe wrapper for flourishcharts bind functions
Removes null bindings to avoid Flourish JavaScript errors}
\usage{
.safe_bind(bind_fn, widget, ...)
}
\description{
Safe wrapper for flourishcharts bind functions
Removes null bindings to avoid Flourish JavaScript errors
}
\keyword{internal}
