% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.split_data_for_display}
\alias{.split_data_for_display}
\title{Split data into main (data rows) and footer (base/summary rows)}
\usage{
.split_data_for_display(x)
}
\description{
Split data into main (data rows) and footer (base/summary rows)
}
\keyword{internal}
