% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{.tab_registry}
\alias{.tab_registry}
\title{Internal registry for tab helpers and statistics}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.tab_registry
}
\description{
Internal registry for tab helpers and statistics
}
\keyword{internal}
