% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_for_chart_type}
\alias{.transform_for_chart_type}
\title{Transform data based on chart type requirements}
\usage{
.transform_for_chart_type(df, chart_type, long = TRUE)
}
\description{
Transform data based on chart type requirements
}
\keyword{internal}
