% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_grouped_bar}
\alias{.transform_grouped_bar}
\title{Transform for grouped bar charts}
\usage{
.transform_grouped_bar(df, n_rows, n_cols, col_names, stat_info)
}
\description{
Transform for grouped bar charts
}
\keyword{internal}
