% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_line_area}
\alias{.transform_line_area}
\title{Transform for line/area charts}
\usage{
.transform_line_area(df, n_rows, n_cols, col_names, stat_info)
}
\description{
Transform for line/area charts
}
\keyword{internal}
