% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_pie_donut}
\alias{.transform_pie_donut}
\title{Transform for pie/donut charts}
\usage{
.transform_pie_donut(df, n_rows, n_cols, col_names)
}
\description{
Transform for pie/donut charts
}
\keyword{internal}
