% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_scatter}
\alias{.transform_scatter}
\title{Transform for scatter plots}
\usage{
.transform_scatter(df, n_rows, n_cols, col_names)
}
\description{
Transform for scatter plots
}
\keyword{internal}
