% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_stacked_bar}
\alias{.transform_stacked_bar}
\title{Transform for stacked bar charts}
\usage{
.transform_stacked_bar(df, n_rows, n_cols, col_names, stat_info)
}
\description{
Transform for stacked bar charts
}
\keyword{internal}
