% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{.transform_wide_format}
\alias{.transform_wide_format}
\title{Transform for wide format charts}
\usage{
.transform_wide_format(df, chart_type)
}
\description{
Transform for wide format charts
}
\keyword{internal}
