% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoreactable.R
\name{.validate_sig_comparison}
\alias{.validate_sig_comparison}
\title{Validate and select significance comparison}
\usage{
.validate_sig_comparison(significance, sig_comparison)
}
\description{
Validate and select significance comparison
}
\keyword{internal}
