% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dsl_contains_value}
\alias{dsl_contains_value}
\title{DSL Contains Value}
\usage{
dsl_contains_value(dsl, value)
}
\arguments{
\item{dsl}{DSL expression}

\item{value}{Value to check}
}
\value{
Logical or NA if not determinable
}
\description{
Checks if DSL expression selects a specific value (where determinable).
}
\keyword{internal}
