% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dsl_get_values}
\alias{dsl_get_values}
\title{DSL Get Values}
\usage{
dsl_get_values(dsl)
}
\arguments{
\item{dsl}{DSL expression}
}
\value{
Vector of values or NULL if not determinable
}
\description{
Extracts all literal values from DSL expression where determinable.
}
