% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dsl_get_variables}
\alias{dsl_get_variables}
\title{DSL Get Variables}
\usage{
dsl_get_variables(dsl)
}
\arguments{
\item{dsl}{DSL expression}
}
\value{
Character vector of variable names
}
\description{
Extracts all variable names from DSL expression.
}
