% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dsl_references_variable}
\alias{dsl_references_variable}
\title{DSL References Variable}
\usage{
dsl_references_variable(dsl, var_name)
}
\arguments{
\item{dsl}{DSL expression}

\item{var_name}{Variable name to check}
}
\value{
Logical
}
\description{
Checks if DSL expression references a specific variable.
}
\keyword{internal}
