% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{eval_dsl}
\alias{eval_dsl}
\title{Evaluate DSL Expression}
\usage{
eval_dsl(expr, data)
}
\arguments{
\item{expr}{DSL expression}

\item{data}{Data frame for evaluation context}
}
\value{
Numeric vector of length nrow(data)
}
\description{
Evaluates a validated DSL expression in the context of data.
}
\keyword{internal}
