% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{exact_match_matcher}
\alias{exact_match_matcher}
\title{Create an exact match matcher}
\usage{
exact_match_matcher(target)
}
\arguments{
\item{target}{Value to match against}
}
\value{
Matcher function
}
\description{
Returns a function that checks for identity with a target value.
}
\examples{
\dontrun{
matcher <- exact_match_matcher(quote(gender == 1))
matcher(quote(gender == 1))  # TRUE
matcher(quote(gender == 2))  # FALSE
}
}
\keyword{internal}
