% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{expr_to_dsl}
\alias{expr_to_dsl}
\title{Convert Expression to DSL}
\usage{
expr_to_dsl(expr, data = NULL, dpdict = NULL)
}
\arguments{
\item{expr}{R expression}

\item{data}{Optional data for context}

\item{dpdict}{Optional dpdict}
}
\value{
Expression if valid DSL, NULL otherwise
}
\description{
Checks if expression is valid DSL.
}
\keyword{internal}
