% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{extract_all_values_from_or_chain}
\alias{extract_all_values_from_or_chain}
\title{Extract All Values from OR Chain}
\usage{
extract_all_values_from_or_chain(expr)
}
\description{
Extract All Values from OR Chain
}
\keyword{internal}
