% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{extract_common_labels_by_question}
\alias{extract_common_labels_by_question}
\title{extract_common_labels_by_question
creates a dataframe with a column for common labels for each question group, given a dpdict.
optionally calls data_edit before returning, for any manual tidying of labels}
\usage{
extract_common_labels_by_question(
  temp_dpdict,
  column_to_extract_from = "variable_labels",
  editfirst = TRUE
)
}
\arguments{
\item{temp_dpdict}{a dpdict}

\item{column_to_extract_from}{column to look for common labels in}

\item{editfirst}{whether to call data_edit before returning, for any manual tidying of labels}
}
\value{
dataframe with a row for each unique question group in temp_dpdict and a column with the lcs from column_to_extract_from for each common question group
}
\description{
extract_common_labels_by_question
creates a dataframe with a column for common labels for each question group, given a dpdict.
optionally calls data_edit before returning, for any manual tidying of labels
}
\examples{
temp_dat <- get_big_test_dat(n=10)
temp_dpdict <- create_dict(temp_dat, prefill = TRUE) # create fresh basic dpdict
temp_dpdict <- temp_dpdict[, !(names(temp_dpdict) \%in\% c("old_variable_names",
                                                   "old_variable_labels",
                                                   "old_value_labels"))]
if(!"question_group" \%in\% names(temp_dpdict)){
 temp_dpdict$question_group <- gsub(paste0("_",".*"), "", temp_dpdict$variable_names)
}
temp_dpdict$question_group <- paste0(temp_dpdict$question_group, "_a")
extract_common_labels_by_question(temp_dpdict, editfirst = FALSE)
}
