% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{extract_common_variable_from_or_chain}
\alias{extract_common_variable_from_or_chain}
\title{Extract Common Variable from OR Chain}
\usage{
extract_common_variable_from_or_chain(expr)
}
\description{
Extract Common Variable from OR Chain
}
\keyword{internal}
