% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{extract_primary_variable}
\alias{extract_primary_variable}
\title{Extract Primary Variable from DSL}
\usage{
extract_primary_variable(dsl)
}
\arguments{
\item{dsl}{DSL expression}
}
\value{
Character variable name or NULL
}
\description{
Gets the main variable from a DSL expression (first encountered).
}
\keyword{internal}
