% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{extract_sort_values_cols}
\alias{extract_sort_values_cols}
\title{Extract values for sorting columns}
\usage{
extract_sort_values_cols(tab_result, by_rows, col_indices, col_names)
}
\description{
Extract values for sorting columns
}
\keyword{internal}
