% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{extract_sort_values_rows}
\alias{extract_sort_values_rows}
\title{Extract values for sorting rows}
\usage{
extract_sort_values_rows(tab_result, by_cols, row_indices)
}
\description{
Extract values for sorting rows
}
\keyword{internal}
