% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{extract_suffix_from_label}
\alias{extract_suffix_from_label}
\title{Extract suffix from a label string}
\usage{
extract_suffix_from_label(label)
}
\description{
Extract suffix from a label string
}
\keyword{internal}
