% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{filter_low_base_cells}
\alias{filter_low_base_cells}
\title{Filter layout grid based on low base thresholds}
\usage{
filter_low_base_cells(layout, store, threshold)
}
\arguments{
\item{layout}{Layout list with grid, row_labels, col_labels, etc}

\item{store}{Cell store}

\item{threshold}{Minimum base count threshold}
}
\description{
Filter layout grid based on low base thresholds
}
\keyword{internal}
