% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{find_collapsible_col_groups_meta}
\alias{find_collapsible_col_groups_meta}
\title{Find groups of columns that can be collapsed based on meta attributes}
\usage{
find_collapsible_col_groups_meta(
  col_labels,
  data_matrix,
  col_arrays,
  common_separators = c(" - ", ": ", " | ", " / "),
  field = NULL
)
}
\arguments{
\item{col_labels}{Character vector of column labels}

\item{data_matrix}{Matrix of data values}

\item{col_arrays}{List of column arrays with meta attributes}

\item{common_separators}{Separators for fallback string matching}

\item{field}{Optional field to collapse by (e.g., "ival", "ivar")}
}
\value{
Named list of column groups
}
\description{
Find groups of columns that can be collapsed based on meta attributes
}
\keyword{internal}
