% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{find_collapsible_col_groups_string}
\alias{find_collapsible_col_groups_string}
\title{Helper to do string-based column grouping (fallback)}
\usage{
find_collapsible_col_groups_string(col_labels, data_matrix, common_separators)
}
\description{
Helper to do string-based column grouping (fallback)
}
\keyword{internal}
