% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{find_collapsible_groups}
\alias{find_collapsible_groups}
\title{Find groups of rows that can be collapsed}
\usage{
find_collapsible_groups(row_labels, data_cols, common_separators)
}
\arguments{
\item{row_labels}{Character vector of row labels}

\item{data_cols}{Data frame of numeric columns}

\item{common_separators}{Separators to check}
}
\value{
List of collapsible groups
}
\description{
Find groups of rows that can be collapsed
}
\keyword{internal}
