% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{find_matching_cells}
\alias{find_matching_cells}
\title{Find cells matching a criterion}
\usage{
find_matching_cells(tab_result, criterion, match_type, dimension)
}
\arguments{
\item{tab_result}{Tab result object}

\item{criterion}{Matching criterion (type depends on .match)}

\item{match_type}{Character: matching strategy}

\item{dimension}{Character: "row" or "col"}
}
\value{
List of matching cell objects
}
\description{
Find cells matching a criterion
}
\keyword{internal}
