% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{flatten_operator}
\alias{flatten_operator}
\title{Flatten Associative Operator}
\usage{
flatten_operator(expr, op)
}
\arguments{
\item{expr}{Expression with associative operator}

\item{op}{Operator to flatten ("&" or "|")}
}
\value{
List of flattened arguments
}
\description{
Converts (a & b) & c to list(a, b, c)
}
\keyword{internal}
