% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{format_cells}
\alias{format_cells}
\title{Functions that control cell value formatting for display.
Operations manipulate metadata; formatting applied at display/export time.
Set global cell formatting options}
\usage{
format_cells(tab_result, digits = NULL, empty_zeros = FALSE, na_display = "NA")
}
\arguments{
\item{tab_result}{A tab_result object}

\item{digits}{Number of decimal places (NULL to use statistic default)}

\item{empty_zeros}{Logical, show empty string for zero values}

\item{na_display}{String to display for NA values}
}
\value{
Modified tab_result with updated formatting metadata
}
\description{
Controls how all cells are formatted for display.
}
\examples{
# Format cells in tab results
dat <- get_basic_test_dat()
result <- tab(dat, labelledordinal, binarycategoricalasfactor) \%>\%
  format_cells(digits = 1, empty_zeros = TRUE)
}
