% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{formula_to_array}
\alias{formula_to_array}
\title{Convert formula specification to numeric array
Convert formula specification to numeric array}
\usage{
formula_to_array(formula_spec, data)
}
\arguments{
\item{formula_spec}{Parsed formula specification}

\item{data}{Data frame}
}
\value{
Numeric vector of length nrow(data)
}
\description{
Convert formula specification to numeric array
Convert formula specification to numeric array
}
\keyword{internal}
