% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{generate_cell_id}
\alias{generate_cell_id}
\title{Generate next sequential cell ID}
\usage{
generate_cell_id(store)
}
\arguments{
\item{store}{A cell_store object}
}
\value{
Character cell ID (e.g., "c_000001")
}
\description{
Generate next sequential cell ID
}
\keyword{internal}
