% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{generate_dimension_labels}
\alias{generate_dimension_labels}
\title{Generate Dimension Labels}
\usage{
generate_dimension_labels(keys, dimension_type, dpdict, reference_var = NULL)
}
\description{
Generate Dimension Labels
}
\keyword{internal}
