% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{get_affix_df}
\alias{get_affix_df}
\title{get_affix_df}
\usage{
get_affix_df(
  full_string,
  affix_type = "prefix",
  seps_priority = c("- ", "_", "."),
  strings_to_remove = c("Selected Choice"),
  prioritise_caps = TRUE,
  filter_results = TRUE,
  sort_results = TRUE
)
}
\arguments{
\item{full_string}{a string, e.g. a variable label}

\item{affix_type}{"suffix" or "prefix"}

\item{seps_priority}{a vector of strings representing separators to check for (in order)}

\item{strings_to_remove}{a vector of any strings to remove before working}

\item{prioritise_caps}{whether to sort final dataframe prioritising capital letters}

\item{filter_results}{whether to remove any rows where the affix is an empty string}

\item{sort_results}{whether to sort the final dataframe}
}
\value{
a dataframe summarising seps found in the full_string, the substrings separated, and their locations
}
\description{
gets a dataframe summarising seps found in the full_string, the substrings separated, and their locations
}
\details{
used by get_unique_suffixes

note that sep_count_from_end/ sep_count_from_start is the count within that sep, not the count across all seps
}
\examples{
get_affix_df("Q1: This is a variable - This is a statement")
}
