% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{get_all_bases}
\alias{get_all_bases}
\title{Get all bases (columnar access)}
\usage{
get_all_bases(store)
}
\arguments{
\item{store}{A cell_store object}
}
\value{
Numeric vector of all bases
}
\description{
Get all bases (columnar access)
}
\keyword{internal}
