% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_all_cells_with_positions}
\alias{get_all_cells_with_positions}
\title{Get All Cells with Positions}
\usage{
get_all_cells_with_positions(tab_result)
}
\description{
Get All Cells with Positions
}
\keyword{internal}
