% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{get_all_values}
\alias{get_all_values}
\title{Get all values (columnar access)}
\usage{
get_all_values(store)
}
\arguments{
\item{store}{A cell_store object}
}
\value{
Numeric vector of all values
}
\description{
Fast access to all cell values using columnar storage.
}
\keyword{internal}
