% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_basic_test_dat}
\alias{get_basic_test_dat}
\title{get_basic_test_dat
creates a basic dataframe meant to replicate survey data for testing
with columns for respid, numeric, oe, boolean, and 3 versions of each of ordinal, categorical and binary categorical - as labelled, factors, and characters}
\usage{
get_basic_test_dat(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
get_basic_test_dat
creates a basic dataframe meant to replicate survey data for testing
with columns for respid, numeric, oe, boolean, and 3 versions of each of ordinal, categorical and binary categorical - as labelled, factors, and characters
}
\examples{
get_basic_test_dat()
}
