% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_big_test_dat}
\alias{get_big_test_dat}
\title{get_big_test_dat
creates a dataframe meant to replicate all common question types in survey data}
\usage{
get_big_test_dat(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
get_big_test_dat
creates a dataframe meant to replicate all common question types in survey data
}
\examples{
get_big_test_dat()
}
