% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_big_test_dat_with_prefixes}
\alias{get_big_test_dat_with_prefixes}
\title{get_big_test_dat_with_prefixes}
\usage{
get_big_test_dat_with_prefixes(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
wrapper around get_big_test_dat. calls get_big_test_dat() then adds variable names up to any seps as prefixes to the variable labels.
}
\examples{
get_big_test_dat_with_prefixes()
}
