% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{get_cell}
\alias{get_cell}
\title{Get a single cell by ID}
\usage{
get_cell(store, cell_id)
}
\arguments{
\item{store}{A cell_store object}

\item{cell_id}{Character cell ID}
}
\value{
Cell object or NULL if not found
}
\description{
Get a single cell by ID
}
\keyword{internal}
