% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{get_cells}
\alias{get_cells}
\title{Get multiple cells by IDs}
\usage{
get_cells(store, cell_ids)
}
\arguments{
\item{store}{A cell_store object}

\item{cell_ids}{Character vector of cell IDs}
}
\value{
List of cell objects
}
\description{
Get multiple cells by IDs
}
\keyword{internal}
