% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{get_derive_operation}
\alias{get_derive_operation}
\title{Get registered derive operation by ID}
\usage{
get_derive_operation(id)
}
\arguments{
\item{id}{Character identifier}
}
\description{
Get registered derive operation by ID
}
\keyword{internal}
