% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{get_display_label}
\alias{get_display_label}
\title{Get display label based on prefix_variable_label setting}
\usage{
get_display_label(
  var_name,
  dpdict = NULL,
  prefix_variable_label = TRUE,
  category_name = NULL
)
}
\arguments{
\item{var_name}{Variable name}

\item{dpdict}{Data dictionary}

\item{prefix_variable_label}{Whether to include variable prefix}

\item{category_name}{Category name for expanded variables (NULL for non-expanded)}
}
\description{
Get display label based on prefix_variable_label setting
}
\keyword{internal}
