% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{get_gate}
\alias{get_gate}
\title{Get registered gate by ID}
\usage{
get_gate(id)
}
\arguments{
\item{id}{Character identifier}
}
\value{
Gate factory function or NULL if not found
}
\description{
Get registered gate by ID
}
\keyword{internal}
