% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{get_helper}
\alias{get_helper}
\title{Get registered helper by ID}
\usage{
get_helper(id)
}
\arguments{
\item{id}{Character identifier}
}
\description{
Get registered helper by ID
}
\keyword{internal}
