% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{get_longest_common_substring}
\alias{get_longest_common_substring}
\title{get_longest_common_substring}
\usage{
get_longest_common_substring(string1, string2, fromstart = FALSE)
}
\arguments{
\item{string1}{a string}

\item{string2}{another string}

\item{fromstart}{boolean. if true, finds longest common substring from the start of string1. if false, looks for lcs at any position.}
}
\value{
substring which represents the longest common starting substring found between string1 and string2, with an attribute lcs_distance representing the number of characters in the lcs
}
\description{
finds the long common substring (lcs) from two given strings
}
\examples{
get_longest_common_substring("boat", "boar")
}
