% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_minimal_unlabelled_test_dat}
\alias{get_minimal_unlabelled_test_dat}
\title{get_minimal_unlabelled_test_dat}
\usage{
get_minimal_unlabelled_test_dat(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
creates a very simple dataframe with columns for uid, csat, and nps
}
\examples{
get_minimal_unlabelled_test_dat()
}
