% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_problematic_metadata_test_dat}
\alias{get_problematic_metadata_test_dat}
\title{get_problematic_metadata_test_dat}
\usage{
get_problematic_metadata_test_dat(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
creates a dataframe with various problematic metadata patterns to test error handling
}
\examples{
get_problematic_metadata_test_dat()
}
