% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatagenerators.R
\name{get_routing_test_dat}
\alias{get_routing_test_dat}
\title{get_routing_test_dat}
\usage{
get_routing_test_dat(n = 100, random_seed = 123)
}
\arguments{
\item{n}{number of rows to create}

\item{random_seed}{for reproducability}
}
\value{
a dataframe
}
\description{
creates a dataframe specifically for testing more complex routing patterns to test metadata handling
}
\examples{
get_routing_test_dat()
}
