% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{get_significance_test}
\alias{get_significance_test}
\title{Get registered significance test by ID}
\usage{
get_significance_test(id)
}
\description{
Get registered significance test by ID
}
\keyword{internal}
