% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{get_statistic}
\alias{get_statistic}
\title{Get registered statistic by ID}
\usage{
get_statistic(id)
}
\arguments{
\item{id}{Character identifier}
}
\description{
Get registered statistic by ID
}
\keyword{internal}
