% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{get_updated_seps}
\alias{get_updated_seps}
\title{get_updated_seps}
\usage{
get_updated_seps(temp_dat, sep_analysis, seps_to_use = NULL)
}
\arguments{
\item{temp_dat}{a survey data frame}

\item{sep_analysis}{the result of check_seps(temp_dat)}

\item{seps_to_use}{NULL by default, else provide a list with elements for 'var name', 'prefix', and 'statement' specifying the sep to use instead of the most common sep}
}
\value{
a data frame summarising old and new (updated) variable names and variable labels
}
\description{
updates separator patterns in a survey data file
}
\details{
by default will update all seps to the most common seps
}
\examples{
temp_dat <- get_big_test_dat_with_prefixes()
get_updated_seps(temp_dat, check_seps(temp_dat))
}
