% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{get_value_labels_for_codes}
\alias{get_value_labels_for_codes}
\title{Get Value Labels for Codes}
\usage{
get_value_labels_for_codes(var_name, values, dpdict = NULL)
}
\arguments{
\item{var_name}{Variable name}

\item{values}{Value codes}

\item{dpdict}{Data dictionary}
}
\value{
Character vector of labels
}
\description{
Looks up value labels from dpdict for given variable and value codes.
}
\keyword{internal}
