% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_var_label}
\alias{get_var_label}
\title{Get variable label from dpdict or use variable name}
\usage{
get_var_label(var_name, dpdict = NULL)
}
\description{
Get variable label from dpdict or use variable name
}
\keyword{internal}
