% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{get_variable_array_type}
\alias{get_variable_array_type}
\title{Determine the type of a variable array for statistics validation}
\usage{
get_variable_array_type(array)
}
\arguments{
\item{array}{Numeric vector representing a variable array}
}
\value{
Character string: "categorical", "numeric", or "other"
}
\description{
Determine the type of a variable array for statistics validation
}
\examples{
get_variable_array_type(c(0, 1, 0, 1, NA))  # "categorical"
get_variable_array_type(c(1.5, 2.3, 4.1))   # "numeric"
get_variable_array_type(c("a", "b"))         # "other"
}
