% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{glue_tabs_cols_cellbased}
\alias{glue_tabs_cols_cellbased}
\title{Glue tabs side-by-side (columns)}
\usage{
glue_tabs_cols_cellbased(tab1, tab2, sep, prefix)
}
\description{
Glue tabs side-by-side (columns)
}
\keyword{internal}
