% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{glue_tabs_rows_cellbased}
\alias{glue_tabs_rows_cellbased}
\title{Glue tabs vertically (rows)}
\usage{
glue_tabs_rows_cellbased(tab1, tab2, sep, prefix)
}
\description{
Glue tabs vertically (rows)
}
\keyword{internal}
