% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{\%+-\%}
\alias{\%+-\%}
\title{Tolerance specification operator}
\usage{
x \%+-\% tolerance
}
\arguments{
\item{x}{Formula object}

\item{tolerance}{Numeric tolerance (scalar or length-2 vector)}
}
\value{
formula_with_tolerance object
}
\description{
Infix operator for attaching tolerances to constraint formulas.
Enables concise specification: (constraint ~ target) \%+-\% tolerance
}
