% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{gsub_by_dict}
\alias{gsub_by_dict}
\title{gsub_by_dict}
\usage{
gsub_by_dict(
  reference_dict,
  string_to_change,
  find_column_name = "find_values",
  replacement_column_name = "replace_values"
)
}
\arguments{
\item{reference_dict}{either a dataframe with columns with names matching find_column_name and replacement_column_name, or a named list}

\item{string_to_change}{the string to apply gsub to}

\item{find_column_name}{required if reference_dict is a data frame: the name of the column to look in for a matching substring}

\item{replacement_column_name}{required if reference_dict is a data frame: the name of the column to find the associated replacement string}
}
\value{
string_to_change, with all replacements made
}
\description{
searches in string_to_change for any value in find_column_name in reference_dict and replaces with the associated replacement_column_name value
for as many matches as found
}
\examples{
gsub_by_dict(list("apple" = "fruit", "carrot" = "vegetable", "rose" = "flower"),
  "I have an apple and a carrot.")
}
