% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtoflourish.R
\name{guess_flourish_chart_type}
\alias{guess_flourish_chart_type}
\title{Guess a sensible Flourish chart_type from a tab_result}
\usage{
guess_flourish_chart_type(tab)
}
\description{
Guess a sensible Flourish chart_type from a tab_result
}
\keyword{internal}
