% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_store.R
\name{has_cell}
\alias{has_cell}
\title{Check if cell ID exists in store}
\usage{
has_cell(store, cell_id)
}
\arguments{
\item{store}{A cell_store object}

\item{cell_id}{Character cell ID}
}
\value{
Logical
}
\description{
Check if cell ID exists in store
}
\keyword{internal}
