% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{hide_rows_except}
\alias{hide_rows_except}
\title{Hide rows that don't match a pattern}
\usage{
hide_rows_except(tab_result, pattern)
}
\arguments{
\item{tab_result}{A tab_result object}

\item{pattern}{Regex pattern - rows matching this will remain visible}
}
\value{
Modified tab_result with non-matching rows hidden
}
\description{
Convenience function to hide all rows except those matching a pattern.
}
