% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{identify_meta_indices}
\alias{identify_meta_indices}
\title{Identify which rows or columns are meta (summary/base) positions}
\usage{
identify_meta_indices(store, grid, dimension = c("row", "col"))
}
\arguments{
\item{store}{A cell_store object}

\item{grid}{Layout grid matrix containing cell IDs}

\item{dimension}{Either "row" or "col"}
}
\value{
Integer vector of indices that are meta positions
}
\description{
Checks cells in the grid to identify which rows/columns contain
summary rows, summary columns, or base rows that should be excluded
from significance testing.
}
\keyword{internal}
