% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{identify_special_cols}
\alias{identify_special_cols}
\title{Identify special columns (summary columns)}
\usage{
identify_special_cols(tab_result, col_names)
}
\description{
Identify special columns (summary columns)
}
\keyword{internal}
