% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{identify_special_rows}
\alias{identify_special_rows}
\title{Identify special rows (Base and summary rows)}
\usage{
identify_special_rows(tab_result)
}
\description{
Identify special rows (Base and summary rows)
}
\keyword{internal}
