% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{insert_into_str}
\alias{insert_into_str}
\title{insert_into_str}
\usage{
insert_into_str(
  string_to_insert_into,
  string_to_insert,
  string_to_find,
  position = "before",
  instances_to_insert_at = "ALL",
  strict_mode = FALSE,
  insert_at_end_if_not_found = TRUE
)
}
\arguments{
\item{string_to_insert_into}{the string to return, complete with insertions}

\item{string_to_insert}{the string to be inserted}

\item{string_to_find}{the string to find to determine the insert location}

\item{position}{before or after the string_to_find}

\item{instances_to_insert_at}{instance(s) of string_to_find to insert next to. by default "ALL" but can be any number or vector of numbers <= number of instances.}

\item{strict_mode}{if TRUE and instance_to_insert_at > the number of instances, results in an error. if FALSE, just inserts for all instances.}

\item{insert_at_end_if_not_found}{when TRUE and string_to_find not found, string_to_insert will be inserted at the end of string_to_insert_into}
}
\value{
a string, string_to_insert_into, with insertions
}
\description{
inserts a string before or after a position in another string, with the position identified by a third string
by default inserts before/after every instance of string_to_find but can optionally specify which instances to insert before/after
}
\examples{
insert_into_str("this is a test string", "new ", "test")

}
