% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{is.survey_data}
\alias{is.survey_data}
\title{is.survey_data}
\usage{
is.survey_data(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if the object is a survey_data object, FALSE otherwise.
}
\description{
Check if an object is a survey_data object
}
\examples{
dat <- get_minimal_labelled_test_dat()
survey_obj <- create_survey_data(dat)
is.survey_data(survey_obj)
}
