% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{is.weighting_constraint}
\alias{is.weighting_constraint}
\title{Type checking for constraint objects}
\usage{
is.weighting_constraint(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a weighting_constraint
}
\description{
Type checking for constraint objects
}
