% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{labelled_case_when}
\alias{labelled_case_when}
\title{Simple labelled case_when with proper class inheritance}
\usage{
labelled_case_when(..., .variable_label = NULL)
}
\arguments{
\item{...}{case_when conditions using lv(value, "label") syntax}

\item{.variable_label}{Variable label to apply}
}
\value{
A properly labelled vector using haven::labelled class
}
\description{
Simple labelled case_when with proper class inheritance
}
