% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{labelled_map_dfc}
\alias{labelled_map_dfc}
\title{labelled_map_dfc}
\usage{
labelled_map_dfc(.x, .f, ...)
}
\arguments{
\item{.x}{dataframe}

\item{.f}{function}

\item{...}{function arguments}
}
\value{
.x transformed to have applied .f to each column per purrr::map_dfc, but with variable labels reapplied at the end
}
\description{
simple wrapper around map_dfc to reapply the original variable labels (these are usually dropped)
}
\examples{
labelled_map_dfc(get_minimal_labelled_test_dat(), function(x) ifelse(is.na(x), 0, x))
}
