% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{list_tab_significance_tests}
\alias{list_tab_significance_tests}
\title{List all registered significance tests}
\usage{
list_tab_significance_tests()
}
\description{
List all registered significance tests
}
