% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{merge_left}
\alias{merge_left}
\title{merge_left}
\usage{
merge_left(x, y, by, overwrite = FALSE)
}
\arguments{
\item{x}{a data frame to join to}

\item{y}{the data frame to join with}

\item{by}{the id by which to join}

\item{overwrite}{boolean. if TRUE, y values will replace x values. if FALSE, y will fill in gaps in x values but not overwrite.}
}
\value{
the joined and merged data frame
}
\description{
a dplyr left_join, but when there is a column in x and y, attempts to merge instead of creating a duplicate column
i.e.
if x already has a value for a given id and is in y, use x (by default, with overwrite == FALSE)
if x id not in y then keep x value
if y id not in x then add y value
optionally with overwrite == TRUE, y values will replace x values
}
\examples{
df1 <- data.frame(id = c(1, 2, 3), x = c("a", "b", "c"), y = c(10, 20, 30))
df2 <- data.frame(id = c(1, 2, 4), x = c("A", "B", "D"), z = c(100, 200, 400))
merge_left(df1, df2, by = "id", overwrite = TRUE)
}
