% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{merge_tab_metadata}
\alias{merge_tab_metadata}
\title{Merge metadata from two tabs}
\usage{
merge_tab_metadata(result, tab1, tab2, direction, prefix, new_col_names = NULL)
}
\description{
Merge metadata from two tabs
}
\keyword{internal}
