% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{mutate.survey_data}
\alias{mutate.survey_data}
\title{Adds or modifies columns in the survey data, automatically updating metadata for new variables
using update_dict_with_metadata. Also preserves appropriate metadata for modified variables.
Supports custom variable labels using the realiselabelled_vec() function.
Automatically preserves value labels for appropriate transformations.}
\usage{
\method{mutate}{survey_data}(.data, ...)
}
\arguments{
\item{.data}{A survey_data object.}

\item{...}{Mutation expressions passed to dplyr::mutate. Variable labels can be specified
using realiselabelled_vec(expr, variable_label="label").}
}
\value{
A new survey_data object with mutated data and updated metadata.
}
\description{
Adds or modifies columns in the survey data, automatically updating metadata for new variables
using update_dict_with_metadata. Also preserves appropriate metadata for modified variables.
Supports custom variable labels using the realiselabelled_vec() function.
Automatically preserves value labels for appropriate transformations.
}
\examples{
# Add a new variable with a custom label
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
survey_obj \%>\%
  dplyr::mutate(new_var = realiselabelled_vec(uid + uid, variable_label = "uid squared"))

# Simple assignment preserves value labels automatically
survey_obj \%>\%
  dplyr::mutate(csat_copy = csat)  # Value labels preserved

# Mathematical operations don't preserve value labels (usually inappropriate)
survey_obj \%>\%
  dplyr::mutate(uid_doubled = uid * 2)  # No value labels preserved
}
