% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{my_view}
\alias{my_view}
\title{Display a Data Frame in RStudio Viewer When Available}
\usage{
my_view(x, title = deparse(substitute(x)))
}
\arguments{
\item{x}{An object (typically a data frame, tibble, or similar) to be viewed.}

\item{title}{An optional title for the viewer tab. Defaults to the name of the object.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
This function displays a data frame or similar object using the RStudio data viewer
if available. If RStudio is not detected, it falls back to the standard \code{utils::View()} function.
}
\details{
This approach avoids issues on macOS where \code{utils::View()} may invoke XQuartz by default.
}
\examples{
\dontrun{
my_view(mtcars)
}
}
