% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{names.tab_result}
\alias{names.tab_result}
\title{Get column names from tab_result}
\usage{
\method{names}{tab_result}(x)
}
\arguments{
\item{x}{A tab_result object}
}
\description{
Get column names from tab_result
}
