% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{natural_order}
\alias{natural_order}
\title{Natural ordering of character vectors}
\usage{
natural_order(x, decreasing = FALSE, na.last = TRUE)
}
\description{
Natural ordering of character vectors
}
\keyword{internal}
