% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{ncol.survey_data}
\alias{ncol.survey_data}
\title{Get the number of columns in a survey_data object}
\usage{
ncol.survey_data(x)
}
\arguments{
\item{x}{A survey_data object}
}
\value{
The number of columns in the data component
}
\description{
Get the number of columns in a survey_data object
}
\examples{
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
ncol(survey_obj)
}
