% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{negate_matcher}
\alias{negate_matcher}
\title{Create a negated matcher}
\usage{
negate_matcher(matcher)
}
\arguments{
\item{matcher}{Matcher function to negate}
}
\value{
Negated matcher function
}
\description{
Returns a function that negates another matcher function.
}
\examples{
\dontrun{
is_summary <- function(x) isTRUE(x)
is_not_summary <- negate_matcher(is_summary)
is_not_summary(FALSE)  # TRUE
is_not_summary(TRUE)   # FALSE
}
}
\keyword{internal}
