% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{new_derive_operation}
\alias{new_derive_operation}
\title{Low-level constructor for derive operations}
\usage{
new_derive_operation(id, processor, description = NULL)
}
\arguments{
\item{id}{Character identifier for the operation}

\item{processor}{Function that processes the operation}

\item{description}{Optional description}
}
\description{
Low-level constructor for derive operations
}
\keyword{internal}
