% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{new_derive_spec}
\alias{new_derive_spec}
\title{Create a derive operation specification}
\usage{
new_derive_spec(op_id, ...)
}
\arguments{
\item{op_id}{Operation identifier}

\item{...}{Parameters for the operation}
}
\description{
Helper to create derive specification objects that can be passed to derive()
or specified in tab() parameters.
}
\keyword{internal}
