% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{new_tab_sig_test}
\alias{new_tab_sig_test}
\title{Low-level constructor for significance tests}
\usage{
new_tab_sig_test(id, processor, name = NULL, description = NULL, ...)
}
\description{
Low-level constructor for significance tests
}
\keyword{internal}
