% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{nrow.survey_data}
\alias{nrow.survey_data}
\title{Get the number of rows in a survey_data object}
\usage{
nrow.survey_data(x)
}
\arguments{
\item{x}{A survey_data object}
}
\value{
The number of rows in the data component
}
\description{
Get the number of rows in a survey_data object
}
\examples{
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
nrow(survey_obj)
}
