% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{null_to_NA}
\alias{null_to_NA}
\title{null_to_NA}
\usage{
null_to_NA(x)
}
\arguments{
\item{x}{a vector to sapply through}
}
\value{
the input list, with all NULLs converted to NAs
}
\description{
accepts a list and converts all NULLs in the list to NAs
}
\examples{
null_to_NA(c(NULL, 1, 2, 3, NULL))
}
