% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{number_to_word_rank}
\alias{number_to_word_rank}
\title{number_to_word_rank}
\usage{
number_to_word_rank(number)
}
\arguments{
\item{number}{a numeric value to convert to a ranking word}
}
\value{
a ranking word, e.g. "first"
}
\description{
accepts a number, returns the respective ranking word up to 50, e.g. 1 returns "first"
}
\examples{
number_to_word_rank(1)
}
