% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{parse_by_parameter}
\alias{parse_by_parameter}
\title{Parse the flexible 'by' parameter into named filter expressions}
\usage{
parse_by_parameter(by, data, dpdict = NULL)
}
\arguments{
\item{by}{The by parameter from multi_tab}

\item{data}{The data frame}

\item{dpdict}{Optional data dictionary}
}
\value{
Named list of filter expressions
}
\description{
Parse the flexible 'by' parameter into named filter expressions
}
\keyword{internal}
