% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{parse_expression_vars}
\alias{parse_expression_vars}
\title{Parse R expression to extract variable constraints}
\usage{
parse_expression_vars(expr)
}
\arguments{
\item{expr}{R expression (quote(...))}
}
\value{
List of variable metadata lists
}
\description{
Analyzes an R expression to identify variable-value relationships.
}
\examples{
\dontrun{
parse_expression_vars(quote(x == 5))
# Returns: list(list(ivar = "x", ival = 5))

parse_expression_vars(quote(x \%in\% c(4, 5)))
# Returns: list(list(ivar = "x", ival = c(4, 5)))

parse_expression_vars(quote(x == 5 & y == 2))
# Returns: list(
#   list(ivar = "x", ival = 5),
#   list(ivar = "y", ival = 2)
# )
}
}
\keyword{internal}
