% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{parse_sort_spec}
\alias{parse_sort_spec}
\title{Parse sort specification into standardized format}
\usage{
parse_sort_spec(sort_spec, tab_result, dimension)
}
\description{
Parse sort specification into standardized format
}
\keyword{internal}
