% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{parse_table_formula}
\alias{parse_table_formula}
\title{Parse table formula expressions using NSE}
\usage{
parse_table_formula(expr, data, dpdict = NULL, helpers = NULL)
}
\arguments{
\item{expr}{An expression captured using rlang::enquo}

\item{data}{The data frame to evaluate against}

\item{dpdict}{Optional data dictionary for metadata}
}
\value{
A list with type, components, and label
}
\description{
Parse table formula expressions using NSE
}
\keyword{internal}
