% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{prepare_eval_data}
\alias{prepare_eval_data}
\title{Prepare data for expression evaluation by converting haven_labelled to numeric}
\usage{
prepare_eval_data(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame with haven_labelled variables converted to numeric
}
\description{
Prepare data for expression evaluation by converting haven_labelled to numeric
}
\keyword{internal}
