% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.formula_with_tolerance}
\alias{print.formula_with_tolerance}
\title{Print method for formula_with_tolerance}
\usage{
\method{print}{formula_with_tolerance}(x, ...)
}
\arguments{
\item{x}{A formula_with_tolerance object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for formula_with_tolerance
}
