% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.summary.unified_weighting_result}
\alias{print.summary.unified_weighting_result}
\title{Print method for summary of unified weighting results}
\usage{
\method{print}{summary.unified_weighting_result}(x, ...)
}
\description{
Print method for summary of unified weighting results
}
