% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{print.survey_data}
\alias{print.survey_data}
\title{Print method for survey_data objects}
\usage{
\method{print}{survey_data}(x, ...)
}
\arguments{
\item{x}{A survey_data object}

\item{...}{Additional arguments passed to print}
}
\value{
x invisibly
}
\description{
Print method for survey_data objects
}
\examples{
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
print(survey_obj)
}
