% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.unified_weighting_result}
\alias{print.unified_weighting_result}
\title{Print method for unified weighting results}
\usage{
\method{print}{unified_weighting_result}(x, ...)
}
\arguments{
\item{x}{A unified_weighting_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for unified weighting results
}
