% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_config.R
\name{print.weighting_config}
\alias{print.weighting_config}
\title{Print method for weighting_config}
\usage{
\method{print}{weighting_config}(x, ...)
}
\arguments{
\item{x}{A weighting_config object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for weighting_config
}
