% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.weighting_constraint}
\alias{print.weighting_constraint}
\title{Print method for constraints}
\usage{
\method{print}{weighting_constraint}(x, ...)
}
\arguments{
\item{x}{A weighting_constraint object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for constraints
}
