% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.weights_infeasible}
\alias{print.weights_infeasible}
\title{Print method for infeasible weighting results}
\usage{
\method{print}{weights_infeasible}(x, ...)
}
\arguments{
\item{x}{weights_infeasible object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for infeasible weighting results
}
