% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{process_delta_vs}
\alias{process_delta_vs}
\title{Process delta_vs derive operation}
\usage{
process_delta_vs(tab_result, from_col, to_col = NULL, label = NULL)
}
\description{
Process delta_vs derive operation
}
\keyword{internal}
