% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{process_helper_for_specs}
\alias{process_helper_for_specs}
\title{Process helper to extract components for spec generators}
\usage{
process_helper_for_specs(formula_spec, data, dpdict = NULL, all_helpers = NULL)
}
\description{
Process helper to extract components for spec generators
}
\keyword{internal}
