% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{process_index_vs}
\alias{process_index_vs}
\title{Process index_vs derive operation}
\usage{
process_index_vs(
  tab_result,
  base_col,
  multiplier = 100,
  label = NULL,
  cols = NULL
)
}
\description{
Process index_vs derive operation
}
\keyword{internal}
