% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{process_share_of_sum}
\alias{process_share_of_sum}
\title{Process share_of_sum derive operation}
\usage{
process_share_of_sum(tab_result, by, label = NULL)
}
\description{
Process share_of_sum derive operation
}
\keyword{internal}
