% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{process_sum_if}
\alias{process_sum_if}
\title{Process sum_if derive operation}
\usage{
process_sum_if(tab_result, metadata_field, dimension, label_fn = NULL)
}
\description{
Process sum_if derive operation
}
\keyword{internal}
