% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{puncts_to_pattern}
\alias{puncts_to_pattern}
\title{puncts_to_pattern}
\usage{
puncts_to_pattern(puncts)
}
\arguments{
\item{puncts}{Character vector of punctuation marks}
}
\value{
A string suitable for use in regex pattern matching
}
\description{
Convert punctuation vector to regex pattern
}
\details{
Takes a vector of punctuation characters or strings (allowing whitespace) and returns a regex-safe pattern
that matches any one of them. Escapes special regex characters and whitespace.
}
