% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{query_dict}
\alias{query_dict}
\title{query_dict}
\usage{
query_dict(dict, key_label, key, value_label, .default = NA)
}
\arguments{
\item{dict}{a data frame which must include key_label and value_label as column names (can also include other names that aren't used)}

\item{key_label}{the name of the data frame column to look in for the key}

\item{key}{the value to look for in the key_label column}

\item{value_label}{the name of the data frame column to look in for the return value}

\item{.default}{the value to return if the key can't be found in the key_label column, or if there if the return value has length 0. NA by default.}
}
\value{
the value(s) in the value_label column in the same row(s) as the key value(s) in the key_label column. (only one key value can be given but if there are multiple instances of this value in the key_label column, all associated return values will be returned)
}
\description{
queries a data frame as if a 'dictionary'
}
\examples{
item_price_lookup <- data.frame(item = c("apple", "carrot"), price = c("£0.20","£0.50"))
query_dict(item_price_lookup, "item", "apple", "price")
}
