% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{register_builtin_derive_operations}
\alias{register_builtin_derive_operations}
\title{Register built-in derive operations}
\usage{
register_builtin_derive_operations()
}
\description{
Called automatically on package load, but can be called explicitly if needed.
}
\keyword{internal}
