% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{remove_empty_rows_cols}
\alias{remove_empty_rows_cols}
\title{Remove empty rows and columns from grid}
\usage{
remove_empty_rows_cols(tab_result)
}
\arguments{
\item{tab_result}{A tab_result object}
}
\value{
Modified tab_result with empty rows/cols removed
}
\description{
Post-processing function that removes rows/columns where all cells are NA.
Also removes corresponding layout_defs.
}
\keyword{internal}
