% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{remove_overlapping_groups}
\alias{remove_overlapping_groups}
\title{Remove overlapping groups, keeping those with most rows}
\usage{
remove_overlapping_groups(groups)
}
\arguments{
\item{groups}{List of groups}
}
\value{
List of non-overlapping groups
}
\description{
Remove overlapping groups, keeping those with most rows
}
\keyword{internal}
