% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{resolve_helper_target}
\alias{resolve_helper_target}
\title{Resolve helper target (variable name, vector of names, or group name)}
\usage{
resolve_helper_target(
  name_or_names,
  data,
  dpdict = NULL,
  error_context = "helper target"
)
}
\description{
Resolve helper target (variable name, vector of names, or group name)
}
\keyword{internal}
