% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{resolve_labels_to_expressions}
\alias{resolve_labels_to_expressions}
\title{Resolve labels to expressions}
\usage{
resolve_labels_to_expressions(labels, layout_defs)
}
\arguments{
\item{labels}{Character vector of labels}

\item{layout_defs}{List of layout_def objects}
}
\value{
List of expressions corresponding to labels
}
\description{
Maps label strings to the expressions they represent by looking up in
layout_defs. Builds efficient lookup table.
}
\keyword{internal}
