% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{resolve_versus_to_column_name}
\alias{resolve_versus_to_column_name}
\title{Resolve versus parameter to actual column name}
\usage{
resolve_versus_to_column_name(versus, tab_result)
}
\description{
Resolve versus parameter to actual column name
}
\keyword{internal}
