% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{rownames.tab_result}
\alias{rownames.tab_result}
\title{Get row names of tab_result}
\usage{
rownames.tab_result(x)
}
\arguments{
\item{x}{A tab_result object}
}
\description{
Get row names of tab_result
}
