% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{rows_list}
\alias{rows_list}
\title{Create named list of row specifications}
\usage{
rows_list(...)
}
\arguments{
\item{...}{Named expressions for row specifications}
}
\value{
List for use in tab()
}
\description{
Create named list of row specifications
}
\examples{
tab(data, rows = rows_list("Total" = q1, "Young" = q1 * (age < 30)))
}
