% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{set_matcher}
\alias{set_matcher}
\title{Create a set membership matcher}
\usage{
set_matcher(valid_set)
}
\arguments{
\item{valid_set}{Vector of valid values}
}
\value{
Matcher function
}
\description{
Returns a function that checks if a value is in a valid set.
}
\examples{
\dontrun{
matcher <- set_matcher(c("column_pct", "row_pct"))
matcher("column_pct")  # TRUE
matcher("mean")        # FALSE
}
}
\keyword{internal}
