% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{show_base}
\alias{show_base}
\title{Show base row or column}
\usage{
show_base(tab_result, orientation = c("auto", "row", "column"))
}
\arguments{
\item{tab_result}{A tab_result object}

\item{orientation}{Where to show base: "auto" (default), "row", or "column"}
}
\value{
Modified tab_result with base added
}
\description{
Adds a base (sample size) row or column to the grid. Computes base
values from existing cells. Orientation is automatically determined
from the base variation pattern, or can be specified explicitly.
}
\examples{
# Show base in tab results
dat <- get_basic_test_dat()
tab(dat, labelledordinal, binarycategoricalasfactor, show_base = FALSE) \%>\%
  show_base()  # Add base back with automatic orientation
  
tab(dat, labelledordinal, binarycategoricalasfactor, show_base = FALSE) \%>\%
  show_base(orientation = "row")  # Force base as row
}
