% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{sort_args_canonical}
\alias{sort_args_canonical}
\title{Sort Arguments in Canonical Order}
\usage{
sort_args_canonical(args)
}
\arguments{
\item{args}{List of expressions}
}
\value{
Sorted list
}
\description{
Sorts expression arguments for normalization (alphabetically by deparse).
}
\keyword{internal}
