% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{sort_tab_dimension}
\alias{sort_tab_dimension}
\title{Internal function to sort a single dimension (rows or cols)}
\usage{
sort_tab_dimension(tab_result, sort_spec, dimension = c("rows", "cols"))
}
\description{
Internal function to sort a single dimension (rows or cols)
}
\keyword{internal}
