% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{specs_equal}
\alias{specs_equal}
\title{Compare two specifications for equality}
\usage{
specs_equal(spec1, spec2)
}
\arguments{
\item{spec1}{First specification expression}

\item{spec2}{Second specification expression}
}
\value{
Logical
}
\description{
Determines if two specification expressions are identical.
Used for exact matching when building layout grid.
}
\keyword{internal}
