% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{standardise_survey_separators}
\alias{standardise_survey_separators}
\title{Standardise separators, in survey data and dpdict}
\usage{
standardise_survey_separators(temp_dat, temp_dpdict = NULL, seps_to_use = NULL)
}
\arguments{
\item{temp_dat}{A data frame containing the survey data}

\item{temp_dpdict}{Optional. The associated dpdict.}

\item{seps}{Optional. A list specifying the separators to use. If NULL, current patterns detected in the data will be used.}
}
\value{
A list with dat and dpdict, both with standardised variable names
}
\description{
Required for update_dict_with_metadata
}
\examples{
dat <- get_minimal_labelled_test_dat()
dpdict <- create_dict(dat, prefill = TRUE)
standardised <- standardise_survey_separators(dat, dpdict)
}
