% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{str.tab_result}
\alias{str.tab_result}
\title{Display structure of tab_result}
\usage{
\method{str}{tab_result}(object, ...)
}
\arguments{
\item{object}{A tab_result object}

\item{...}{Passed to the underlying method}
}
\description{
Display structure of tab_result
}
