% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{strsplit_keep_delim}
\alias{strsplit_keep_delim}
\title{strsplit_keep_delim}
\usage{
strsplit_keep_delim(string_to_split, split)
}
\arguments{
\item{string_to_split}{the string to split}

\item{split}{the delimiter to split by}
}
\value{
a vector of strings
}
\description{
splits a string by a given delimiter and returns as a flat vector of strings, including the delimiters
i.e. like strsplit, but adds the delimiters back in
allows delimiter to be of length 0, e.g. "", in which case would return with no splits, but will error if string_to_split of length 0
}
\examples{
strsplit_keep_delim("a-b-c", "-")
}
