% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{[.survey_data}
\alias{[.survey_data}
\title{Subsetting method for survey_data objects}
\usage{
\method{[}{survey_data}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A survey_data object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{drop}{If TRUE, returns a vector when only one column is selected}
}
\value{
A new survey_data object containing only the selected data and corresponding metadata
}
\description{
Subsetting method for survey_data objects
}
\examples{
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
# Select first 3 rows
subset_rows <- survey_obj[1:3, ]
# Select specific columns
subset_cols <- survey_obj[, c("uid", "csat")]
}
