% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{[[.survey_data}
\alias{[[.survey_data}
\title{Single column extraction method for survey_data objects}
\usage{
\method{[[}{survey_data}(x, i)
}
\arguments{
\item{x}{A survey_data object}

\item{i}{Name or index of the column to extract}
}
\value{
The selected column from the data
}
\description{
Single column extraction method for survey_data objects
}
\examples{
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
# Extract a column by name
uid_vector <- survey_obj[["uid"]]
# Extract a column by position
first_column <- survey_obj[[1]]
}
