% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{[[.tab_result}
\alias{[[.tab_result}
\title{Extract column or cell from tab_result}
\usage{
\method{[[}{tab_result}(x, i)
}
\arguments{
\item{x}{A tab_result object}

\item{i}{Element name or index}
}
\description{
Extract column or cell from tab_result
}
