% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{sub_char_by_index}
\alias{sub_char_by_index}
\title{sub_char_by_index}
\usage{
sub_char_by_index(string, index, replacement)
}
\arguments{
\item{string}{the string to replace within}

\item{index}{the index of the character(s) in the given string to replace}

\item{replacement}{the replacement character(s)}
}
\value{
the string with the character(s) at the given index replaced
}
\description{
accepts a string and replaces just the character(s) at a given index with a specified string
index must be a single numeric or range of consecutive numerics
if indexes given a range spanning multiple character positions, replaces with only one instance of the replacement string
replacement string can be "", removing the character(s) at the given index
}
\examples{
sub_char_by_index("testing", 3, "x")
}
