% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{summary.unified_weighting_result}
\alias{summary.unified_weighting_result}
\title{Summary method for unified weighting results}
\usage{
\method{summary}{unified_weighting_result}(object, ...)
}
\arguments{
\item{object}{A unified_weighting_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for unified weighting results
}
