% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{summary.weights_infeasible}
\alias{summary.weights_infeasible}
\title{Summary method for infeasible weighting results}
\usage{
\method{summary}{weights_infeasible}(object, ...)
}
\arguments{
\item{object}{weights_infeasible object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for infeasible weighting results
}
