% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydatar-package.R
\docType{package}
\name{surveydatar-package}
\alias{surveydatar}
\alias{surveydatar-package}
\title{surveydatar: Tools for data processing survey data in a market research setting}
\description{
Functions to support the full data processing workflow for a typical market research project using survey data, from ingestion to basic formatting to the efficient creation of new variables.
}
\author{
\strong{Maintainer}: Matt Harris \email{mdharris89@gmail.com}

}
\keyword{internal}
