% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{tail.tab_result}
\alias{tail.tab_result}
\title{Get last n rows of tab_result}
\usage{
\method{tail}{tab_result}(x, n = 6L, ...)
}
\arguments{
\item{x}{A tab_result object}

\item{n}{Number of rows to return (default 6)}

\item{...}{Passed to the underlying method}
}
\description{
Get last n rows of tab_result
}
