% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{top_box}
\alias{top_box}
\title{Select top n response options}
\usage{
top_box(var, n = 1)
}
\arguments{
\item{var}{Variable name or expression}

\item{n}{Number of top options to select (default 1)}
}
\value{
Expression for use in tab()
}
\description{
Select top n response options
}
\examples{
top_box(satisfaction, 2)
}
