% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{update_aliases}
\alias{update_aliases}
\title{update_aliases}
\usage{
update_aliases(temp_dpdict, questions_dict = NULL)
}
\arguments{
\item{temp_dpdict}{dpdict, e.g. from create_dict_with_metadata}

\item{questions_dict}{optional. a questions_dict that must have columns for question_group and question_alias
and the same number of values as there are unique question_groups in temp_dpdict}
}
\value{
temp_dpdict with alias_with_suffix always updated, and question_alias updated if questions_dict was provided
}
\description{
update_aliases
}
\examples{
update_aliases(create_dict_with_metadata(get_big_test_dat()))
}
