% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{update_dat_from_dpdict}
\alias{update_dat_from_dpdict}
\title{update_dat_from_dpdict
convenience function for updating all variable names, variable labels and value labels in a dat according to a dpdict}
\usage{
update_dat_from_dpdict(temp_dat, temp_dpdict)
}
\arguments{
\item{temp_dat}{survey data dataframe}

\item{temp_dpdict}{dpdict related to temp_dat}
}
\value{
version of temp_dat with variable names, variable labels and values labels updated according to temp_dpdict
}
\description{
update_dat_from_dpdict
convenience function for updating all variable names, variable labels and value labels in a dat according to a dpdict
}
\examples{
temp_dat <- get_minimal_labelled_test_dat()
temp_dpdict <- data.frame(
  old_variable_names = c("uid", "csat", "nps_response"),
  old_variable_labels = c("uid", "Satisfaction", "NPS response"),
  variable_names = c("uid", "customer_satisfaction", "nps_rating"),
  variable_labels = c("uid", "Customer Satisfaction Score", "NPS rating (0-10)"),
  stringsAsFactors = FALSE
)
}
