% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessingfunctions.R
\name{update_labelled_values}
\alias{update_labelled_values}
\title{update_labelled_values}
\usage{
update_labelled_values(
  x,
  old_value,
  new_value,
  new_label = NULL,
  rows_to_ignore = NULL,
  force_add_new_label = FALSE,
  noisy = FALSE
)
}
\arguments{
\item{x}{A labelled vector (anything for which sjlabelled::is_labelled() is TRUE)}

\item{old_value}{Value to be replaced (can be NA)}

\item{new_value}{Value to replace old_value with}

\item{new_label}{Label to attach to new_value (NULL or "" if no label desired)}

\item{rows_to_ignore}{Logical vector of same length as x indicating which rows to leave unchanged (NULL = none)}

\item{force_add_new_label}{Logical. If old_value not found, should label for new_value still be added/updated?}

\item{noisy}{Logical. If TRUE, prints progress notes}
}
\value{
A labelled vector with both values and labels attribute updated
}
\description{
replaces a value in a labelled vector and updates/adds value-labels to keep metadata consistent
}
\examples{
# Create a labelled vector
v <- sjlabelled::set_labels(c(1, NA, 2, NA), labels = c("Yes" = 1, "Maybe" = 2))

# Replace NA with 0 and label it
v2 <- update_labelled_values(v, NA, 0, "No / missing")
sjlabelled::get_labels(v2)

# Change specific rows only
rows_to_skip <- c(TRUE, TRUE, FALSE, FALSE)
v3 <- update_labelled_values(v2, 0, 9, "Definitely maybe", rows_to_skip)
sjlabelled::get_labels(v3)
}
