% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{validate_dat_dpdict_alignment}
\alias{validate_dat_dpdict_alignment}
\title{validate_dat_dpdict_alignment}
\usage{
validate_dat_dpdict_alignment(temp_dat, temp_dpdict, warn_only = FALSE)
}
\arguments{
\item{temp_dat}{survey data dataframe}

\item{temp_dpdict}{'dpdict' to check for alignment with temp_dat}

\item{warn_only}{logical. if TRUE returns warnings instead of errors}
}
\value{
(invisibly) TRUE if passes all checks, else warnings or errors
}
\description{
simple checks that variable names and labels in dat and dpdict match and are aligned
}
\examples{
temp_dat <- get_minimal_labelled_test_dat()
temp_dpdict <- create_dict(temp_dat)
validate_dat_dpdict_alignment(temp_dat, temp_dpdict)
}
