% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{validate_filter_expression}
\alias{validate_filter_expression}
\title{Validate that filter expression references valid variables}
\usage{
validate_filter_expression(expr, data)
}
\description{
Validate that filter expression references valid variables
}
\keyword{internal}
