% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{validate_no_dpdict_duplicates}
\alias{validate_no_dpdict_duplicates}
\title{validate_no_dpdict_duplicates}
\usage{
validate_no_dpdict_duplicates(
  temp_dpdict,
  check_variable_names = TRUE,
  check_variable_labels = TRUE,
  check_alias_with_suffix = TRUE,
  ignore_variable_name_from_label = FALSE,
  variable_name_sep = ": ",
  warn_only = FALSE
)
}
\arguments{
\item{temp_dpdict}{a dataframe, assumed to be survey data}

\item{check_variable_names}{boolean. whether to check for and report on duplicate variable names.}

\item{check_variable_labels}{boolean. whether to check for and report on duplicate variable labels.}

\item{check_alias_with_suffix}{boolean. whether to check for and report on duplicate aliases with suffixes.}

\item{ignore_variable_name_from_label}{if TRUE, removes any cases of paste0(new_variable_name, variable_name_sep) from new_variable_label before working
for example, would remove "SC1: ", from "SC1: Country"}

\item{variable_name_sep}{specify the sep used to delineate variable name from label, e.g. ":" in "SC1: Country"}

\item{warn_only}{boolean. if TRUE returns warnings instead of errors}
}
\value{
TRUE (invisibly) if no issues, else warnings or errors
}
\description{
checks for unique values in variable names and labels, and, if the column exists, alias_with_suffix, in dpdict, reporting back with a message listing any duplicates
}
\details{
returns(invisibly) a temp_dpdict filtered for duplicates, so it can be more easily viewed e.g. with View()
}
\examples{
validate_no_dpdict_duplicates(create_dict_with_metadata(get_big_test_dat(n=10)))
}
