% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{validate_survey_data}
\alias{validate_survey_data}
\title{validate_survey_data}
\usage{
validate_survey_data(x)
}
\arguments{
\item{x}{A survey_data object to validate.}
}
\value{
TRUE if the object is valid, otherwise it throws an error with a description of the problem.
}
\description{
Validate a survey_data object
}
\examples{
dat <- get_minimal_labelled_test_dat()
survey_obj <- create_survey_data(dat)
validate_survey_data(survey_obj)
}
