% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{value_from_labels}
\alias{value_from_labels}
\title{value_from_labels}
\usage{
value_from_labels(labelled_var, label_to_find)
}
\arguments{
\item{labelled_var}{a vector with value labels}

\item{label_to_find}{string corresponding to one of the value labels for labelled_var}
}
\value{
the value associated with the label_to_find
}
\description{
given a vector with value labels, returns the value associated with the label_to_find
returns NA if label not found
}
\examples{
labelled_test_dat <- get_minimal_labelled_test_dat()
value_from_labels(labelled_test_dat$csat, "Neutral")
}
