% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{with_label}
\alias{with_label}
\title{Create a labeled value for use within mutate.survey_data}
\usage{
with_label(x, label)
}
\arguments{
\item{x}{The value expression}

\item{label}{The variable label to apply}
}
\value{
The value with an attribute indicating the desired label
}
\description{
Create a labeled value for use within mutate.survey_data
}
\examples{
# Used within mutate.survey_data for labeling new variables
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
result <- dplyr::mutate(survey_obj, new_var = with_label(uid * 2, "Doubled ID"))
}
