% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{any_positive}
\alias{any_positive}
\title{Check if any variable in a group is positive}
\usage{
any_positive(vars)
}
\arguments{
\item{vars}{Vector of variable names or single question group name}
}
\value{
A tab_helper object for use in tab() rows/cols/filter
}
\description{
Check if any variable in a group is positive
}
