% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{create_dict_with_metadata}
\alias{create_dict_with_metadata}
\title{create_dict_with_metadata
creates a dpdict with additional metadata fields for question group, variable class and checks for single variable question, dichotomous variable, variable with value labels, and multiresponse variable, which are in turn used to define questiontype
these are all calculated automatically based on function arguments
then additionally creates fields for question_alias, question_description, question_suffix, alias_with_suffix, and question_folder
question_alias is intended as a user-defined identifier for the question but takes the values of question_group by default
question_description, question_suffix, and alias_with_suffix are useful when creating tables and visualisations
suffixes are identified using question_lcs
requires: split_into_question_groups()
most arguments are settings for split_into_question_groups()}
\usage{
create_dict_with_metadata(
  temp_dat,
  variable_name_sep = "_",
  splitbyclass = TRUE,
  splitbynumlabelledvalues = TRUE,
  splitbynoncontiguous = TRUE,
  splitbycommonlabel = TRUE,
  findlongest = TRUE,
  variable_compare_mode = "reduced",
  min_common_strings = 3,
  labelsep = " - ",
  edit_aliases = FALSE,
  noisy = 0
)
}
\arguments{
\item{temp_dat}{survey data dataframe}

\item{variable_name_sep}{"\emph{" by default. string that separates each variable within a common question, e.g. if Q1_1, Q1_2 "}" is the correct variable_name_sep.}

\item{splitbyclass}{boolean. if true, every successive new class is given a new unique suffix}

\item{splitbynumlabelledvalues}{boolean. if true, each variable with a different number of labelled values is given a new unique suffix (because we would expect multiple variables in the same question to have the same number of options, e.g. battery of likert scales)}

\item{splitbynoncontiguous}{boolean. if true, variables with the same variable name prefix that are not located adjacent to each other are given a new unique suffix}

\item{splitbycommonlabel}{boolean. if true, each variable for which the longest common string within the variable label for that variable does not match the longest common string within the question group as a whole, is given a new suffix}

\item{findlongest}{boolean. if true, applies splitbycommonlabel by looking for the longest common substring. (this is computationally expensive.) if false, takes substring preceding labelsep as the commonlabel.}

\item{variable_compare_mode}{either "complete" or "reduced'. setting for when findlongest == TRUE. "complete" is computationally cheaper. if "complete", compares within the entire question group to find lcs for each statement. if "reduced", only looks at the variables before and after the variable in question, according to consecutive_range (a parameter in split_into_question_groups)}

\item{min_common_strings}{integer. setting for when findlongest == TRUE. looks for substrings common to at least min_common_strings other variables in the question group and will take from this subset the longest string.}

\item{labelsep}{string. setting for when findlongest == FALSE, in which case uses the substring in the variable name prior to labelsep as commonlabel.}

\item{edit_aliases}{boolean. whether to allow user to manually edit question_aliases using data_edit before creating alias_with_suffix. (question_folder can be defined at the same time)}

\item{noisy}{integer between 1 and 4. 1 by default. if 1, signals start of attempts to find question groups, and multiresponse. if noisy == 2 provides timing and updates within question groups and multiresponse hunts. noisy > 2 has settings within function for finding question groups.}
}
\value{
a tibble with metadata for temp_dat.
}
\description{
notes:
\itemize{
\item identifies categorical and multiresponse ('select all') variables by the presence of value labels. i.e. if a variable has class labelled but not value labels, it will not be given questiontype multiresponse, categorical or categorical array. (usually it will be a numeric or multinumeric)
\item findlongest=TRUE, with splitbycommonlabel=TRUE takes significantly longer (as it attempts to split by analysing strings). may not be necessary for some processing - if correct common label is always the string prior to label_sep.
\item variable labels should be hierarchically structured, with unique statements at the end. e.g. "Q1. Question text - Attitude statement - answer code"
\item checking for multiresponse variables can also takes a while but is required.
\item if last variable in a question in returned dict just keeps its full string, min_common_strings may be set too high
}
}
\examples{
create_dict_with_metadata(get_big_test_dat(n=10))
}
