% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{datamap}
\alias{datamap}
\title{datamap}
\usage{
datamap(x, view_or_return = "view")
}
\arguments{
\item{x}{A dataframe or survey_data object}

\item{view_or_return}{"view" to call View, or "return" to return the summary as a dataframe}
}
\value{
Calls View by default, or returns the summary dataframe if view_or_return is "return"
}
\description{
creates a dataframe with variable names, labels and other useful info as rows (SPSS style)
}
\details{
useful for quickly understanding structure of a large dataset
note: order of unique values is in order they are found, not necessarily corresponding to the order in which they have in labelling
}
\examples{
datamap(get_basic_test_dat(100), view_or_return = "return")
survey_obj <- create_survey_data(get_basic_test_dat(100))
datamap(survey_obj, view_or_return = "return")
}
