% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{.onLoad}
\alias{.onLoad}
\title{Register built-in helpers and statistics}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name (unused)}

\item{pkgname}{Package name (unused)}
}
\description{
Register built-in helpers and statistics
}
\keyword{internal}
