% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{get_questions_dict}
\alias{get_questions_dict}
\title{get_questions_dict
very simple function to get a questions_dict from a dpdict}
\usage{
get_questions_dict(temp_dpdict)
}
\arguments{
\item{temp_dpdict}{a dpdict with columns for question-level metadata, e.g. questiontype, question_alias}
}
\value{
a questions_dict (subset of dpdict with one row per unique question group)
}
\description{
get_questions_dict
very simple function to get a questions_dict from a dpdict
}
\examples{
get_questions_dict(create_dict_with_metadata(get_big_test_dat()))
}
