% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins.R
\name{index_to}
\alias{index_to}
\title{Index values relative to a base}
\usage{
index_to(var, base = NULL)
}
\arguments{
\item{var}{Variable name or expression}

\item{base}{Base for indexing (default is total)}
}
\value{
Expression for use in tab()
}
\description{
Index values relative to a base
}
