% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{lv}
\alias{lv}
\title{Helper to create labeled value}
\usage{
lv(value, label)
}
\arguments{
\item{value}{The value to assign}

\item{label}{The label for this value}
}
\value{
The original \code{value} with an additional \code{label} attribute.
}
\description{
Helper to create labeled value
}
