% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{mutate.survey_data}
\alias{mutate.survey_data}
\title{Adds or modifies columns in the survey data, automatically updating metadata for new variables
using update_dict_with_metadata. Also preserves appropriate metadata for modified variables.
Supports custom variable labels using the with_label() function.}
\usage{
\method{mutate}{survey_data}(.data, ...)
}
\arguments{
\item{.data}{A survey_data object.}

\item{...}{Mutation expressions passed to dplyr::mutate. Variable labels can be specified
using with_label(expr, "label").}
}
\value{
A new survey_data object with mutated data and updated metadata.
}
\description{
Adds or modifies columns in the survey data, automatically updating metadata for new variables
using update_dict_with_metadata. Also preserves appropriate metadata for modified variables.
Supports custom variable labels using the with_label() function.
}
\examples{
# Add a new variable with a custom label
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
survey_obj \%>\%
  dplyr::mutate(new_var = with_label(survey_obj$dat$uid + survey_obj$dat$uid, "uid squared"))

# Modify an existing variable with a new label
survey_obj \%>\%
  dplyr::mutate(uid = with_label(uid * 2, "uid doubled"))
}
