% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_constructors.R
\name{new_tab_helper}
\alias{new_tab_helper}
\title{Low-level constructor for tab helpers}
\usage{
new_tab_helper(id, processor, ...)
}
\arguments{
\item{id}{Character identifier for the helper}

\item{processor}{Function that processes the helper}

\item{...}{Additional attributes}
}
\description{
Low-level constructor for tab helpers
}
\keyword{internal}
