% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{row_def_matches_expression}
\alias{row_def_matches_expression}
\title{Check if row_def matches an expression}
\usage{
row_def_matches_expression(row_def, expr, row_dsl = NULL)
}
\arguments{
\item{row_def}{A layout_def object}

\item{expr}{Expression to check}

\item{row_dsl}{Optional DSL object from cell's specification (for pivoted grids)}
}
\value{
Logical
}
\description{
Check if row_def matches an expression
}
\keyword{internal}
