% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{split_grid_labels}
\alias{split_grid_labels}
\title{split_grid_labels
for expanding the question alias in a dpdict for grid questions, e.g. brand perceptions where a series of statements are rated for each of a series of brands
e.g. moving from question_alias == 'Q1. Brand Image', question_suffix == 'Brand 1 - Statement 1'...
to question_alias == 'Q1. Brand Image - Brand 1', question_suffix == 'Statement 1'
so that the common question alias is specific to each brand.
Does so based on an example 'stem' (e.g. 'Brand 1') and a count_before_repeat, i.e. the number of statements asked for each brand.
(Assumes the same number of statements exists for each brand.)}
\usage{
split_grid_labels(
  temp_dpdict,
  alias_to_split,
  example_stem_to_add,
  count_before_repeat,
  sep = " - "
)
}
\arguments{
\item{temp_dpdict}{a 'dpdict' dataframe}

\item{alias_to_split}{the question_alias in the temp_dpdict to change}

\item{example_stem_to_add}{as an example, the first stem currently in question_suffix that should be moved to question_alias}

\item{count_before_repeat}{the number of times the desired question_suffix after the stem has been moved repeats before the next stem}

\item{sep}{the sep used in the question_suffix}
}
\value{
an updated dpdict
}
\description{
split_grid_labels
for expanding the question alias in a dpdict for grid questions, e.g. brand perceptions where a series of statements are rated for each of a series of brands
e.g. moving from question_alias == 'Q1. Brand Image', question_suffix == 'Brand 1 - Statement 1'...
to question_alias == 'Q1. Brand Image - Brand 1', question_suffix == 'Statement 1'
so that the common question alias is specific to each brand.
Does so based on an example 'stem' (e.g. 'Brand 1') and a count_before_repeat, i.e. the number of statements asked for each brand.
(Assumes the same number of statements exists for each brand.)
}
\examples{
split_grid_labels(create_dict_with_metadata(get_big_test_dat(n=10)),
                  "labelledmultiresponsegrid_a", "statement 1 ", 2)
}
