% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_methods.R
\name{[.tab_result}
\alias{[.tab_result}
\title{Subset tab_result (materializes first)}
\usage{
\method{[}{tab_result}(x, i, j, drop = FALSE)
}
\description{
Subset tab_result (materializes first)
}
