% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydatar-package.R
\docType{package}
\name{surveydatar-package}
\alias{surveydatar}
\alias{surveydatar-package}
\title{surveydatar: Tools for data processing for survey data in a market research setting}
\description{
Aimed at covering the full data processing workflow for a typical market research project using survey data, from ingestion to basic formatting to the efficient creation of new variables.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
