% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydatar-package.R
\docType{package}
\name{surveydatar-package}
\alias{surveydatar}
\alias{surveydatar-package}
\title{surveydatar: Tools for Processing and Analysing Market Research Survey Data}
\description{
Tools for processing and analysing survey data for market research, including metadata management via the survey_data and dpdict framework and flexible cross-tabulation with tab().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdharris89/surveydatar}
  \item Report bugs at \url{https://github.com/mdharris89/surveydatar/issues}
}

}
\author{
\strong{Maintainer}: Matt Harris \email{mdharris89@gmail.com}

}
\keyword{internal}
