% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymetadatafunctions.R
\name{update_dat_from_dpdict}
\alias{update_dat_from_dpdict}
\title{update_dat_from_dpdict}
\usage{
update_dat_from_dpdict(x, temp_dpdict = NULL)
}
\arguments{
\item{x}{survey data object, or a survey data dataframe (a 'dat')}

\item{temp_dpdict}{dpdict related to the survey_obj or temp_dat}
}
\value{
version of survey_obj or temp_dat with variable names, variable labels and values labels updated according to temp_dpdict
}
\description{
convenience function for updating all variable names, variable labels and value labels in a dat according to a dpdict
}
\examples{
# Using separate temp_dat and temp_dpdict
temp_dat <- get_minimal_labelled_test_dat()
temp_dpdict <- create_dict(temp_dat)
temp_dpdict$variable_names[1] <- "new_uid"
updated_dat <- update_dat_from_dpdict(temp_dat, temp_dpdict)

# Using a survey_data object and a modified temp_dpdict
survey_obj <- create_survey_data(get_minimal_labelled_test_dat())
temp_dpdict <- survey_obj$dpdict
temp_dpdict$variable_names[1] <- "new_uid"
updated_survey_obj <- update_dat_from_dpdict(survey_obj, temp_dpdict)
}
